@@ -6,17 +6,18 @@
   "type": "module",
   "scripts": {
     "start": "node server.js",
-    "dev": "nodemon server.js",
-    "test": "jest --coverage",
-    "test:watch": "jest --watch",
+    "dev": "nodemon --exec node --experimental-modules server.js",
+    "dev:legacy": "nodemon server.js --experimental-modules --es-module-specifier-resolution=node",
+    "test": "node --experimental-modules --es-module-specifier-resolution=node node_modules/.bin/jest --coverage",
+    "test:watch": "node --experimental-modules --es-module-specifier-resolution=node node_modules/.bin/jest --watch",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "format": "prettier --write .",
     "build": "npm run build:clean && npm run build:copy",
     "build:clean": "rm -rf dist",
     "build:copy": "mkdir -p dist && cp -r src/* dist/ && cp package.json dist/",
-    "seed": "node src/utils/seedDatabase.js",
-    "migrate": "node src/utils/migrateDatabase.js",
+    "seed": "node --experimental-modules --es-module-specifier-resolution=node src/utils/seedDatabase.js",
+    "migrate": "node --experimental-modules --es-module-specifier-resolution=node src/utils/migrateDatabase.js",
     "heroku-postbuild": "npm run build"
   },
   "keywords": [
@@ -30,26 +31,33 @@
   "author": "Your Name",
   "license": "MIT",
   "dependencies": {
-    "all": "^0.0.0",
     "bcryptjs": "^2.4.3",
     "compression": "^1.8.1",
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "dotenv": "^16.6.1",
+    "ejs": "^3.1.10",
     "express": "^4.22.1",
     "express-mongo-sanitize": "^2.2.0",
-    "express-rate-limit": "^6.7.0",
+    "express-rate-limit": "^7.2.0",
     "express-validator": "^7.3.1",
     "helmet": "^7.2.0",
+    "ioredis": "^5.8.2",
     "joi": "^17.9.2",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^7.0.3",
     "morgan": "^1.10.1",
     "multer": "^1.4.5-lts.1",
+    "node-cache": "^5.1.2",
     "nodemailer": "^6.9.1",
+    "rate-limit-redis": "^4.3.1",
+    "rate-limiter-flexible": "^9.0.0",
+    "sanitize-html": "^2.17.0",
     "socket.io": "^4.6.1",
     "stripe": "^12.1.1",
     "validator": "^13.9.0",
+    "winston": "^3.19.0",
+    "xss": "^1.0.15",
     "xss-clean": "^0.1.4"
   },
   "devDependencies": {
@@ -60,12 +68,12 @@
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
     "jest": "^29.5.0",
-    "nodemon": "^2.0.22",
+    "nodemon": "^3.1.9",
     "prettier": "^2.8.7",
     "supertest": "^6.3.3"
   },
   "engines": {
-    "node": ">=16.0.0",
+    "node": ">=18.0.0",
     "npm": ">=8.0.0"
   },
   "repository": {
