@@ -1,24 +1,24 @@
 // server/src/controllers/contactController.js
-const Contact = require('../models/Contact');
-const User = require('../models/User');
-const Notification = require('../models/Notification');
-const ActivityLog = require('../models/ActivityLog');
-const asyncHandler = require('../middleware/asyncHandler');
-const ErrorResponse = require('../utils/errorResponse');
-const sendEmail = require('../utils/emailService');
+import Contact from "../models/Contact.js";
+import User from "../models/User.js";
+import Notification from "../models/Notification.js";
+import ActivityLog from "../models/ActivityLog.js";
+import asyncHandler from "../middleware/asyncHandler.js";
+import ErrorResponse from "../utils/errorResponse.js";
+import { sendEmail } from "../utils/emailService.js";
 
 // @desc    Submit contact form (Public)
 // @route   POST /api/contact
 // @access  Public
-exports.submitContact = asyncHandler(async (req, res, next) => {
+const submitContact = asyncHandler(async (req, res, next) => {
   const {
     name,
     email,
     phone,
     subject,
     message,
-    category = 'general',
-    bloodGroup = 'not_applicable',
+    category = "general",
+    bloodGroup = "not_applicable",
     district,
     upazila,
     consentGiven = false,
@@ -27,30 +27,39 @@ exports.submitContact = asyncHandler(async (req, res, next) => {
 
   // Validate required fields
   if (!name || !email || !subject || !message) {
-    return next(new ErrorResponse('Name, email, subject, and message are required', 400));
+    return next(
+      new ErrorResponse("Name, email, subject, and message are required", 400)
+    );
   }
 
   // Validate message length
   if (message.length < 10) {
-    return next(new ErrorResponse('Message must be at least 10 characters', 400));
+    return next(
+      new ErrorResponse("Message must be at least 10 characters", 400)
+    );
   }
 
   // Validate email format
   const emailRegex = /^\S+@\S+\.\S+$/;
   if (!emailRegex.test(email)) {
-    return next(new ErrorResponse('Please enter a valid email address', 400));
+    return next(new ErrorResponse("Please enter a valid email address", 400));
   }
 
   // Check for duplicate submissions (same email and similar message within 24 hours)
   const twentyFourHoursAgo = new Date(Date.now() - 24 * 60 * 60 * 1000);
   const duplicate = await Contact.findOne({
     email: email.toLowerCase(),
-    message: { $regex: message.substring(0, 50), $options: 'i' },
+    message: { $regex: message.substring(0, 50), $options: "i" },
     createdAt: { $gte: twentyFourHoursAgo },
   });
 
   if (duplicate) {
-    return next(new ErrorResponse('You have already submitted a similar message recently. Please wait 24 hours before submitting again.', 429));
+    return next(
+      new ErrorResponse(
+        "You have already submitted a similar message recently. Please wait 24 hours before submitting again.",
+        429
+      )
+    );
   }
 
   // Get user if logged in
@@ -60,18 +69,18 @@ exports.submitContact = asyncHandler(async (req, res, next) => {
   }
 
   // Determine priority based on category
-  let priority = 'medium';
-  if (category === 'emergency' || category === 'complaint') {
-    priority = 'high';
-  } else if (category === 'general' || category === 'feedback') {
-    priority = 'low';
+  let priority = "medium";
+  if (category === "emergency" || category === "complaint") {
+    priority = "high";
+  } else if (category === "general" || category === "feedback") {
+    priority = "low";
   }
 
   // Create contact submission
   const contact = await Contact.create({
     name,
     email: email.toLowerCase(),
-    phone: phone || '',
+    phone: phone || "",
     user,
     subject,
     message,
@@ -79,12 +88,12 @@ exports.submitContact = asyncHandler(async (req, res, next) => {
     priority,
     bloodGroup: bloodGroup.toUpperCase(),
     location: {
-      district: district || '',
-      upazila: upazila || '',
+      district: district || "",
+      upazila: upazila || "",
     },
-    userAgent: req.headers['user-agent'] || '',
-    ipAddress: req.ip || '',
-    source: 'contact_form',
+    userAgent: req.headers["user-agent"] || "",
+    ipAddress: req.ip || "",
+    source: "contact_form",
     consentGiven,
     allowMarketing,
     isSubscribed: true,
@@ -95,36 +104,36 @@ exports.submitContact = asyncHandler(async (req, res, next) => {
     await sendEmail({
       to: email,
       subject: `Thank you for contacting Blood Donation App - ${subject}`,
-      template: 'contact-auto-reply',
-      context: {
+      template: "contact-auto-reply",
+      data: { // Changed from context to data
         name,
         subject,
         message,
         category,
         priority,
         referenceId: contact._id.toString().slice(-8),
-        estimatedResponseTime: '24-48 hours',
-        contactEmail: process.env.SUPPORT_EMAIL || 'support@blooddonation.app',
+        estimatedResponseTime: "24-48 hours",
+        contactEmail: process.env.SUPPORT_EMAIL || "support@blooddonation.app",
       },
     });
   } catch (emailError) {
-    console.error('Auto-reply email failed:', emailError);
+    console.error("Auto-reply email failed:", emailError);
     // Don't fail the request if email fails
   }
 
   // Create notification for admins
   const admins = await User.find({
-    role: 'admin',
-    status: 'active',
+    role: "admin",
+    status: "active",
   });
 
-  const adminNotifications = admins.map(admin => ({
+  const adminNotifications = admins.map((admin) => ({
     recipient: admin._id,
     recipientEmail: admin.email,
-    title: 'New Contact Submission 📧',
+    title: "New Contact Submission 📧",
     message: `New ${category} submission from ${name}: ${subject}`,
-    type: 'info',
-    category: 'contact',
+    type: "info",
+    category: "contact",
     priority,
     actionUrl: `/dashboard/contacts/${contact._id}`,
     data: {
@@ -140,25 +149,25 @@ exports.submitContact = asyncHandler(async (req, res, next) => {
   }
 
   // If it's an emergency, also notify volunteers in the same area
-  if (category === 'emergency' && district) {
+  if (category === "emergency" && district) {
     const volunteers = await User.find({
-      role: 'volunteer',
-      status: 'active',
+      role: "volunteer",
+      status: "active",
       district: district,
     });
 
-    const volunteerNotifications = volunteers.map(volunteer => ({
+    const volunteerNotifications = volunteers.map((volunteer) => ({
       recipient: volunteer._id,
       recipientEmail: volunteer.email,
-      title: 'Emergency Contact Submission 🚨',
+      title: "Emergency Contact Submission 🚨",
       message: `Emergency contact from ${name} in ${district}: ${subject}`,
-      type: 'alert',
-      category: 'contact',
-      priority: 'critical',
+      type: "alert",
+      category: "contact",
+      priority: "critical",
       actionUrl: `/dashboard/contacts/${contact._id}`,
       data: {
         contactId: contact._id,
-        category: 'emergency',
+        category: "emergency",
         location: district,
         from: name,
       },
@@ -174,31 +183,31 @@ exports.submitContact = asyncHandler(async (req, res, next) => {
     user: user || null,
     userName: name,
     userEmail: email,
-    userRole: user ? req.user.role : 'anonymous',
-    action: 'Submitted Contact Form',
-    actionType: 'create',
-    category: 'contact',
-    entityType: 'contact',
+    userRole: user ? req.user.role : "anonymous",
+    action: "Submitted Contact Form",
+    actionType: "create",
+    category: "contact",
+    entityType: "contact",
     entityId: contact._id,
     entityName: name,
     description: `Submitted contact form: ${subject}`,
     details: `Category: ${category}, Priority: ${priority}`,
-    status: 'success',
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   res.status(201).json({
     success: true,
     data: contact,
-    message: 'Thank you for contacting us. We will get back to you soon.',
+    message: "Thank you for contacting us. We will get back to you soon.",
   });
 });
 
 // @desc    Get all contacts (Admin/Volunteer only)
 // @route   GET /api/contacts
 // @access  Private/Admin/Volunteer
-exports.getAllContacts = asyncHandler(async (req, res, next) => {
+const getAllContacts = asyncHandler(async (req, res, next) => {
   const page = parseInt(req.query.page, 10) || 1;
   const limit = parseInt(req.query.limit, 10) || 20;
   const skip = (page - 1) * limit;
@@ -207,12 +216,9 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
   const filter = {};
 
   // Role-based filtering
-  if (req.user.role === 'volunteer') {
+  if (req.user.role === "volunteer") {
     // Volunteers see contacts assigned to them or unassigned
-    filter.$or = [
-      { assignedTo: req.user.id },
-      { assignedTo: null },
-    ];
+    filter.$or = [{ assignedTo: req.user.id }, { assignedTo: null }];
   }
 
   // Additional filters
@@ -220,37 +226,40 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
   if (req.query.category) filter.category = req.query.category;
   if (req.query.priority) filter.priority = req.query.priority;
   if (req.query.assignedTo) {
-    if (req.query.assignedTo === 'unassigned') {
+    if (req.query.assignedTo === "unassigned") {
       filter.assignedTo = null;
-    } else if (req.query.assignedTo === 'me') {
+    } else if (req.query.assignedTo === "me") {
       filter.assignedTo = req.user.id;
     } else {
       filter.assignedTo = req.query.assignedTo;
     }
   }
-  
+
   // Search filter
   if (req.query.search) {
     filter.$or = [
-      { name: { $regex: req.query.search, $options: 'i' } },
-      { email: { $regex: req.query.search, $options: 'i' } },
-      { subject: { $regex: req.query.search, $options: 'i' } },
-      { message: { $regex: req.query.search, $options: 'i' } },
+      { name: { $regex: req.query.search, $options: "i" } },
+      { email: { $regex: req.query.search, $options: "i" } },
+      { subject: { $regex: req.query.search, $options: "i" } },
+      { message: { $regex: req.query.search, $options: "i" } },
     ];
   }
 
   // Date filtering
   if (req.query.startDate || req.query.endDate) {
     filter.createdAt = {};
-    if (req.query.startDate) filter.createdAt.$gte = new Date(req.query.startDate);
+    if (req.query.startDate)
+      filter.createdAt.$gte = new Date(req.query.startDate);
     if (req.query.endDate) filter.createdAt.$lte = new Date(req.query.endDate);
   }
 
   // Sorting
   const sort = {};
   if (req.query.sort) {
-    const sortField = req.query.sort.startsWith('-') ? req.query.sort.substring(1) : req.query.sort;
-    sort[sortField] = req.query.sort.startsWith('-') ? -1 : 1;
+    const sortField = req.query.sort.startsWith("-")
+      ? req.query.sort.substring(1)
+      : req.query.sort;
+    sort[sortField] = req.query.sort.startsWith("-") ? -1 : 1;
   } else {
     // Default: priority first, then newest
     sort.priority = -1;
@@ -259,9 +268,9 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
 
   const [contacts, total] = await Promise.all([
     Contact.find(filter)
-      .populate('user', 'name email avatar')
-      .populate('assignedTo', 'name email avatar')
-      .populate('responses.responder', 'name email avatar')
+      .populate("user", "name email avatar")
+      .populate("assignedTo", "name email avatar")
+      .populate("responses.responder", "name email avatar")
       .skip(skip)
       .limit(limit)
       .sort(sort),
@@ -278,10 +287,10 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
         statusDistribution: [
           {
             $group: {
-              _id: '$status',
+              _id: "$status",
               count: { $sum: 1 },
               urgentCount: {
-                $sum: { $cond: [{ $eq: ['$priority', 'urgent'] }, 1, 0] },
+                $sum: { $cond: [{ $eq: ["$priority", "urgent"] }, 1, 0] },
               },
             },
           },
@@ -289,7 +298,7 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
         categoryDistribution: [
           {
             $group: {
-              _id: '$category',
+              _id: "$category",
               count: { $sum: 1 },
             },
           },
@@ -298,7 +307,7 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
         priorityDistribution: [
           {
             $group: {
-              _id: '$priority',
+              _id: "$priority",
               count: { $sum: 1 },
             },
           },
@@ -307,7 +316,11 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
           {
             $group: {
               _id: {
-                $cond: [{ $eq: ['$assignedTo', null] }, 'unassigned', 'assigned'],
+                $cond: [
+                  { $eq: ["$assignedTo", null] },
+                  "unassigned",
+                  "assigned",
+                ],
               },
               count: { $sum: 1 },
             },
@@ -316,21 +329,21 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
         responseTimeStats: [
           {
             $match: {
-              status: { $in: ['resolved', 'closed'] },
-              'responses.0': { $exists: true },
+              status: { $in: ["resolved", "closed"] },
+              "responses.0": { $exists: true },
             },
           },
           {
             $project: {
               createdAt: 1,
-              firstResponse: { $arrayElemAt: ['$responses', 0] },
+              firstResponse: { $arrayElemAt: ["$responses", 0] },
             },
           },
           {
             $project: {
               responseTime: {
                 $divide: [
-                  { $subtract: ['$firstResponse.sentAt', '$createdAt'] },
+                  { $subtract: ["$firstResponse.sentAt", "$createdAt"] },
                   1000 * 60 * 60, // Convert to hours
                 ],
               },
@@ -339,9 +352,9 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
           {
             $group: {
               _id: null,
-              avgHours: { $avg: '$responseTime' },
-              minHours: { $min: '$responseTime' },
-              maxHours: { $max: '$responseTime' },
+              avgHours: { $avg: "$responseTime" },
+              minHours: { $min: "$responseTime" },
+              maxHours: { $max: "$responseTime" },
             },
           },
         ],
@@ -368,30 +381,36 @@ exports.getAllContacts = asyncHandler(async (req, res, next) => {
 // @desc    Get single contact
 // @route   GET /api/contacts/:id
 // @access  Private/Admin/Volunteer
-exports.getContact = asyncHandler(async (req, res, next) => {
+const getContact = asyncHandler(async (req, res, next) => {
   const contact = await Contact.findById(req.params.id)
-    .populate('user', 'name email avatar phone bloodGroup district upazila')
-    .populate('assignedTo', 'name email avatar')
-    .populate('responses.responder', 'name email avatar role')
-    .populate('archivedBy', 'name email');
+    .populate("user", "name email avatar phone bloodGroup district upazila")
+    .populate("assignedTo", "name email avatar")
+    .populate("responses.responder", "name email avatar role")
+    .populate("archivedBy", "name email");
 
   if (!contact) {
-    return next(new ErrorResponse(`Contact not found with id ${req.params.id}`, 404));
+    return next(
+      new ErrorResponse(`Contact not found with id ${req.params.id}`, 404)
+    );
   }
 
   // Check authorization
-  const canView = 
-    req.user.role === 'admin' || 
+  const canView =
+    req.user.role === "admin" ||
     contact.assignedTo?.toString() === req.user.id ||
-    (!contact.assignedTo && req.user.role === 'volunteer');
+    (!contact.assignedTo && req.user.role === "volunteer");
 
   if (!canView) {
-    return next(new ErrorResponse('Not authorized to view this contact', 403));
+    return next(new ErrorResponse("Not authorized to view this contact", 403));
   }
 
   // Mark as read if it was new
-  if (contact.status === 'new' && (req.user.role === 'admin' || contact.assignedTo?.toString() === req.user.id)) {
-    contact.status = 'read';
+  if (
+    contact.status === "new" &&
+    (req.user.role === "admin" ||
+      contact.assignedTo?.toString() === req.user.id)
+  ) {
+    contact.status = "read";
     await contact.save();
   }
 
@@ -404,27 +423,42 @@ exports.getContact = asyncHandler(async (req, res, next) => {
 // @desc    Update contact status
 // @route   PATCH /api/contacts/:id/status
 // @access  Private/Admin/Volunteer
-exports.updateContactStatus = asyncHandler(async (req, res, next) => {
+const updateContactStatus = asyncHandler(async (req, res, next) => {
   const { status, note } = req.body;
-  const validStatuses = ['new', 'read', 'in-progress', 'resolved', 'closed', 'spam'];
+  const validStatuses = [
+    "new",
+    "read",
+    "in-progress",
+    "resolved",
+    "closed",
+    "spam",
+  ];
 
   if (!validStatuses.includes(status)) {
-    return next(new ErrorResponse(`Invalid status. Valid statuses: ${validStatuses.join(', ')}`, 400));
+    return next(
+      new ErrorResponse(
+        `Invalid status. Valid statuses: ${validStatuses.join(", ")}`,
+        400
+      )
+    );
   }
 
   const contact = await Contact.findById(req.params.id);
 
   if (!contact) {
-    return next(new ErrorResponse(`Contact not found with id ${req.params.id}`, 404));
+    return next(
+      new ErrorResponse(`Contact not found with id ${req.params.id}`, 404)
+    );
   }
 
   // Check authorization
-  const canUpdate = 
-    req.user.role === 'admin' || 
-    contact.assignedTo?.toString() === req.user.id;
+  const canUpdate =
+    req.user.role === "admin" || contact.assignedTo?.toString() === req.user.id;
 
   if (!canUpdate) {
-    return next(new ErrorResponse('Not authorized to update this contact', 403));
+    return next(
+      new ErrorResponse("Not authorized to update this contact", 403)
+    );
   }
 
   const oldStatus = contact.status;
@@ -436,7 +470,7 @@ exports.updateContactStatus = asyncHandler(async (req, res, next) => {
   }
 
   // Set follow-up date if status is in-progress
-  if (status === 'in-progress' && !contact.followUpDate) {
+  if (status === "in-progress" && !contact.followUpDate) {
     const followUpDate = new Date();
     followUpDate.setDate(followUpDate.getDate() + 3); // Follow up in 3 days
     contact.followUpDate = followUpDate;
@@ -450,29 +484,31 @@ exports.updateContactStatus = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Updated Contact Status',
-    actionType: 'update',
-    category: 'contact',
-    entityType: 'contact',
+    action: "Updated Contact Status",
+    actionType: "update",
+    category: "contact",
+    entityType: "contact",
     entityId: contact._id,
     entityName: contact.name,
     description: `Changed status from ${oldStatus} to ${status} for contact from ${contact.name}`,
-    details: note || 'No additional note',
-    status: 'success',
+    details: note || "No additional note",
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   // Notify user if status is resolved or closed
-  if ((status === 'resolved' || status === 'closed') && contact.user) {
+  if ((status === "resolved" || status === "closed") && contact.user) {
     await Notification.createSystemNotification({
       recipient: contact.user,
       recipientEmail: contact.email,
-      title: `Contact ${status === 'resolved' ? 'Resolved' : 'Closed'}`,
-      message: `Your contact submission "${contact.subject}" has been ${status}. ${note || 'Thank you for reaching out to us.'}`,
-      type: status === 'resolved' ? 'success' : 'info',
-      category: 'contact',
-      priority: 'medium',
+      title: `Contact ${status === "resolved" ? "Resolved" : "Closed"}`,
+      message: `Your contact submission "${
+        contact.subject
+      }" has been ${status}. ${note || "Thank you for reaching out to us."}`,
+      type: status === "resolved" ? "success" : "info",
+      category: "contact",
+      priority: "medium",
       actionUrl: `/contact/${contact._id}`,
       data: {
         contactId: contact._id,
@@ -495,31 +531,38 @@ exports.updateContactStatus = asyncHandler(async (req, res, next) => {
 // @desc    Assign contact to user
 // @route   PATCH /api/contacts/:id/assign
 // @access  Private/Admin
-exports.assignContact = asyncHandler(async (req, res, next) => {
+const assignContact = asyncHandler(async (req, res, next) => {
   const { assignTo } = req.body;
 
   if (!assignTo) {
-    return next(new ErrorResponse('User ID to assign is required', 400));
+    return next(new ErrorResponse("User ID to assign is required", 400));
   }
 
   const contact = await Contact.findById(req.params.id);
 
   if (!contact) {
-    return next(new ErrorResponse(`Contact not found with id ${req.params.id}`, 404));
+    return next(
+      new ErrorResponse(`Contact not found with id ${req.params.id}`, 404)
+    );
   }
 
   // Check if assignTo user exists and is admin/volunteer
   const assignUser = await User.findById(assignTo);
-  if (!assignUser || (assignUser.role !== 'admin' && assignUser.role !== 'volunteer')) {
-    return next(new ErrorResponse('Can only assign to admins or volunteers', 400));
+  if (
+    !assignUser ||
+    (assignUser.role !== "admin" && assignUser.role !== "volunteer")
+  ) {
+    return next(
+      new ErrorResponse("Can only assign to admins or volunteers", 400)
+    );
   }
 
   const oldAssignee = contact.assignedTo;
   contact.assignedTo = assignTo;
-  
+
   // Update status if it was new
-  if (contact.status === 'new') {
-    contact.status = 'in-progress';
+  if (contact.status === "new") {
+    contact.status = "in-progress";
   }
 
   await contact.save();
@@ -530,28 +573,30 @@ exports.assignContact = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Assigned Contact',
-    actionType: 'assign',
-    category: 'contact',
-    entityType: 'contact',
+    action: "Assigned Contact",
+    actionType: "assign",
+    category: "contact",
+    entityType: "contact",
     entityId: contact._id,
     entityName: contact.name,
     description: `Assigned contact from ${contact.name} to ${assignUser.name}`,
-    details: `Previous assignee: ${oldAssignee ? 'Someone else' : 'Unassigned'}`,
-    status: 'success',
+    details: `Previous assignee: ${
+      oldAssignee ? "Someone else" : "Unassigned"
+    }`,
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   // Notify the assigned user
   await Notification.createSystemNotification({
     recipient: assignTo,
     recipientEmail: assignUser.email,
-    title: 'New Contact Assigned to You 📋',
+    title: "New Contact Assigned to You 📋",
     message: `You have been assigned a contact from ${contact.name}: "${contact.subject}"`,
-    type: 'info',
-    category: 'contact',
-    priority: contact.priority === 'urgent' ? 'high' : 'medium',
+    type: "info",
+    category: "contact",
+    priority: contact.priority === "urgent" ? "high" : "medium",
     actionUrl: `/dashboard/contacts/${contact._id}`,
     data: {
       contactId: contact._id,
@@ -575,26 +620,29 @@ exports.assignContact = asyncHandler(async (req, res, next) => {
 // @desc    Respond to contact
 // @route   POST /api/contacts/:id/respond
 // @access  Private/Admin/Volunteer
-exports.respondToContact = asyncHandler(async (req, res, next) => {
-  const { message, sendVia = 'email', markAsResolved = false } = req.body;
+const respondToContact = asyncHandler(async (req, res, next) => {
+  const { message, sendVia = "email", markAsResolved = false } = req.body;
 
   if (!message) {
-    return next(new ErrorResponse('Response message is required', 400));
+    return next(new ErrorResponse("Response message is required", 400));
   }
 
   const contact = await Contact.findById(req.params.id);
 
   if (!contact) {
-    return next(new ErrorResponse(`Contact not found with id ${req.params.id}`, 404));
+    return next(
+      new ErrorResponse(`Contact not found with id ${req.params.id}`, 404)
+    );
   }
 
   // Check authorization
-  const canRespond = 
-    req.user.role === 'admin' || 
-    contact.assignedTo?.toString() === req.user.id;
+  const canRespond =
+    req.user.role === "admin" || contact.assignedTo?.toString() === req.user.id;
 
   if (!canRespond) {
-    return next(new ErrorResponse('Not authorized to respond to this contact', 403));
+    return next(
+      new ErrorResponse("Not authorized to respond to this contact", 403)
+    );
   }
 
   // Add response
@@ -607,24 +655,25 @@ exports.respondToContact = asyncHandler(async (req, res, next) => {
   });
 
   // Send email response if requested
-  if (sendVia === 'email') {
+  if (sendVia === "email") {
     try {
       await sendEmail({
         to: contact.email,
         subject: `Re: ${contact.subject}`,
-        template: 'contact-response',
-        context: {
+        template: "contact-response",
+        data: { // Changed from context to data
           name: contact.name,
           subject: contact.subject,
           response: message,
           responderName: req.user.name,
           responderRole: req.user.role,
           contactId: contact._id.toString().slice(-8),
-          contactEmail: process.env.SUPPORT_EMAIL || 'support@blooddonation.app',
+          contactEmail:
+            process.env.SUPPORT_EMAIL || "support@blooddonation.app",
         },
       });
     } catch (emailError) {
-      console.error('Response email failed:', emailError);
+      console.error("Response email failed:", emailError);
       // Don't fail the request if email fails
     }
   }
@@ -640,63 +689,66 @@ exports.respondToContact = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Responded to Contact',
-    actionType: 'update',
-    category: 'contact',
-    entityType: 'contact',
+    action: "Responded to Contact",
+    actionType: "update",
+    category: "contact",
+    entityType: "contact",
     entityId: contact._id,
     entityName: contact.name,
     description: `Responded to contact from ${contact.name}`,
     details: `Response sent via ${sendVia}`,
-    status: 'success',
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   res.status(200).json({
     success: true,
     data: contact,
-    message: 'Response sent successfully',
+    message: "Response sent successfully",
   });
 });
 
 // @desc    Set follow-up date
 // @route   PATCH /api/contacts/:id/follow-up
 // @access  Private/Admin/Volunteer
-exports.setFollowUp = asyncHandler(async (req, res, next) => {
+const setFollowUp = asyncHandler(async (req, res, next) => {
   const { followUpDate, note } = req.body;
 
   if (!followUpDate) {
-    return next(new ErrorResponse('Follow-up date is required', 400));
+    return next(new ErrorResponse("Follow-up date is required", 400));
   }
 
   const date = new Date(followUpDate);
   if (date < new Date()) {
-    return next(new ErrorResponse('Follow-up date cannot be in the past', 400));
+    return next(new ErrorResponse("Follow-up date cannot be in the past", 400));
   }
 
   const contact = await Contact.findById(req.params.id);
 
   if (!contact) {
-    return next(new ErrorResponse(`Contact not found with id ${req.params.id}`, 404));
+    return next(
+      new ErrorResponse(`Contact not found with id ${req.params.id}`, 404)
+    );
   }
 
   // Check authorization
-  const canUpdate = 
-    req.user.role === 'admin' || 
-    contact.assignedTo?.toString() === req.user.id;
+  const canUpdate =
+    req.user.role === "admin" || contact.assignedTo?.toString() === req.user.id;
 
   if (!canUpdate) {
-    return next(new ErrorResponse('Not authorized to update this contact', 403));
+    return next(
+      new ErrorResponse("Not authorized to update this contact", 403)
+    );
   }
 
   const oldFollowUpDate = contact.followUpDate;
   contact.followUpDate = date;
-  contact.followUpNotes = note || contact.followUpNotes || '';
+  contact.followUpNotes = note || contact.followUpNotes || "";
 
   // Update status if it was resolved/closed
-  if (contact.status === 'resolved' || contact.status === 'closed') {
-    contact.status = 'in-progress';
+  if (contact.status === "resolved" || contact.status === "closed") {
+    contact.status = "in-progress";
   }
 
   await contact.save();
@@ -707,17 +759,17 @@ exports.setFollowUp = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Set Follow-up Date',
-    actionType: 'update',
-    category: 'contact',
-    entityType: 'contact',
+    action: "Set Follow-up Date",
+    actionType: "update",
+    category: "contact",
+    entityType: "contact",
     entityId: contact._id,
     entityName: contact.name,
     description: `Set follow-up date for contact from ${contact.name}`,
-    details: `Date: ${date.toLocaleDateString()}, Note: ${note || 'No note'}`,
-    status: 'success',
+    details: `Date: ${date.toLocaleDateString()}, Note: ${note || "No note"}`,
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   // Schedule reminder notification
@@ -727,11 +779,11 @@ exports.setFollowUp = asyncHandler(async (req, res, next) => {
   await Notification.create({
     recipient: req.user._id,
     recipientEmail: req.user.email,
-    title: 'Follow-up Reminder 🔔',
+    title: "Follow-up Reminder 🔔",
     message: `Follow up with ${contact.name} about "${contact.subject}"`,
-    type: 'reminder',
-    category: 'contact',
-    priority: 'medium',
+    type: "reminder",
+    category: "contact",
+    priority: "medium",
     scheduledFor: reminderDate,
     actionUrl: `/dashboard/contacts/${contact._id}`,
     data: {
@@ -752,14 +804,16 @@ exports.setFollowUp = asyncHandler(async (req, res, next) => {
 // @desc    Mark contact as spam
 // @route   PATCH /api/contacts/:id/spam
 // @access  Private/Admin
-exports.markAsSpam = asyncHandler(async (req, res, next) => {
+const markAsSpam = asyncHandler(async (req, res, next) => {
   const contact = await Contact.findById(req.params.id);
 
   if (!contact) {
-    return next(new ErrorResponse(`Contact not found with id ${req.params.id}`, 404));
+    return next(
+      new ErrorResponse(`Contact not found with id ${req.params.id}`, 404)
+    );
   }
 
-  contact.status = 'spam';
+  contact.status = "spam";
   await contact.save();
 
   // Log activity
@@ -768,39 +822,43 @@ exports.markAsSpam = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Marked Contact as Spam',
-    actionType: 'update',
-    category: 'contact',
-    entityType: 'contact',
+    action: "Marked Contact as Spam",
+    actionType: "update",
+    category: "contact",
+    entityType: "contact",
     entityId: contact._id,
     entityName: contact.name,
     description: `Marked contact from ${contact.name} as spam`,
     details: `Subject: ${contact.subject}`,
-    status: 'success',
-    severity: 'warning',
+    status: "success",
+    severity: "warning",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   res.status(200).json({
     success: true,
     data: contact,
-    message: 'Contact marked as spam',
+    message: "Contact marked as spam",
   });
 });
 
 // @desc    Archive contact
 // @route   PATCH /api/contacts/:id/archive
 // @access  Private/Admin
-exports.archiveContact = asyncHandler(async (req, res, next) => {
+const archiveContact = asyncHandler(async (req, res, next) => {
   const contact = await Contact.findById(req.params.id);
 
   if (!contact) {
-    return next(new ErrorResponse(`Contact not found with id ${req.params.id}`, 404));
+    return next(
+      new ErrorResponse(`Contact not found with id ${req.params.id}`, 404)
+    );
   }
 
-  if (contact.status !== 'closed' && contact.status !== 'resolved') {
-    return next(new ErrorResponse('Only resolved or closed contacts can be archived', 400));
+  if (contact.status !== "closed" && contact.status !== "resolved") {
+    return next(
+      new ErrorResponse("Only resolved or closed contacts can be archived", 400)
+    );
   }
 
   await contact.archive(req.user.id);
@@ -811,39 +869,46 @@ exports.archiveContact = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Archived Contact',
-    actionType: 'update',
-    category: 'contact',
-    entityType: 'contact',
+    action: "Archived Contact",
+    actionType: "update",
+    category: "contact",
+    entityType: "contact",
     entityId: contact._id,
     entityName: contact.name,
     description: `Archived contact from ${contact.name}`,
-    status: 'success',
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   res.status(200).json({
     success: true,
     data: contact,
-    message: 'Contact archived successfully',
+    message: "Contact archived successfully",
   });
 });
 
 // @desc    Delete contact (Admin only)
 // @route   DELETE /api/contacts/:id
 // @access  Private/Admin
-exports.deleteContact = asyncHandler(async (req, res, next) => {
+const deleteContact = asyncHandler(async (req, res, next) => {
   const contact = await Contact.findById(req.params.id);
 
   if (!contact) {
-    return next(new ErrorResponse(`Contact not found with id ${req.params.id}`, 404));
+    return next(
+      new ErrorResponse(`Contact not found with id ${req.params.id}`, 404)
+    );
   }
 
   // Only allow deletion of spam or very old contacts
   const thirtyDaysAgo = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
-  if (contact.status !== 'spam' && contact.createdAt > thirtyDaysAgo) {
-    return next(new ErrorResponse('Can only delete spam contacts or contacts older than 30 days', 400));
+  if (contact.status !== "spam" && contact.createdAt > thirtyDaysAgo) {
+    return next(
+      new ErrorResponse(
+        "Can only delete spam contacts or contacts older than 30 days",
+        400
+      )
+    );
   }
 
   await contact.deleteOne();
@@ -854,60 +919,59 @@ exports.deleteContact = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Deleted Contact',
-    actionType: 'delete',
-    category: 'contact',
-    entityType: 'contact',
+    action: "Deleted Contact",
+    actionType: "delete",
+    category: "contact",
+    entityType: "contact",
     entityId: contact._id,
     entityName: contact.name,
     description: `Deleted contact from ${contact.name}`,
     details: `Status was: ${contact.status}`,
-    status: 'success',
-    severity: 'warning',
+    status: "success",
+    severity: "warning",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   res.status(200).json({
     success: true,
     data: {},
-    message: 'Contact deleted successfully',
+    message: "Contact deleted successfully",
   });
 });
 
 // @desc    Get contacts needing follow-up
 // @route   GET /api/contacts/needs-follow-up
 // @access  Private/Admin/Volunteer
-exports.getContactsNeedingFollowUp = asyncHandler(async (req, res, next) => {
-  const filter = {
-    followUpDate: { $lte: new Date() },
-    status: { $in: ['new', 'in-progress', 'read'] },
-  };
+const getContactsNeedingFollowUp = asyncHandler(
+  async (req, res, next) => {
+    const filter = {
+      followUpDate: { $lte: new Date() },
+      status: { $in: ["new", "in-progress", "read"] },
+    };
 
-  // Role-based filtering
-  if (req.user.role === 'volunteer') {
-    filter.$or = [
-      { assignedTo: req.user.id },
-      { assignedTo: null },
-    ];
+    // Role-based filtering
+    if (req.user.role === "volunteer") {
+      filter.$or = [{ assignedTo: req.user.id }, { assignedTo: null }];
+    }
+
+    const contacts = await Contact.find(filter)
+      .populate("assignedTo", "name email")
+      .sort({ followUpDate: 1 })
+      .limit(50);
+
+    res.status(200).json({
+      success: true,
+      count: contacts.length,
+      data: contacts,
+    });
   }
-
-  const contacts = await Contact.find(filter)
-    .populate('assignedTo', 'name email')
-    .sort({ followUpDate: 1 })
-    .limit(50);
-
-  res.status(200).json({
-    success: true,
-    count: contacts.length,
-    data: contacts,
-  });
-});
+);
 
 // @desc    Get contact statistics
 // @route   GET /api/contacts/stats
 // @access  Private/Admin
-exports.getContactStats = asyncHandler(async (req, res, next) => {
+const getContactStats = asyncHandler(async (req, res, next) => {
   const today = new Date();
   const startOfToday = new Date(today.setHours(0, 0, 0, 0));
   const startOfWeek = new Date(today.setDate(today.getDate() - today.getDay()));
@@ -922,11 +986,17 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
             $group: {
               _id: null,
               total: { $sum: 1 },
-              new: { $sum: { $cond: [{ $eq: ['$status', 'new'] }, 1, 0] } },
-              inProgress: { $sum: { $cond: [{ $eq: ['$status', 'in-progress'] }, 1, 0] } },
-              resolved: { $sum: { $cond: [{ $eq: ['$status', 'resolved'] }, 1, 0] } },
-              closed: { $sum: { $cond: [{ $eq: ['$status', 'closed'] }, 1, 0] } },
-              spam: { $sum: { $cond: [{ $eq: ['$status', 'spam'] }, 1, 0] } },
+              new: { $sum: { $cond: [{ $eq: ["$status", "new"] }, 1, 0] } },
+              inProgress: {
+                $sum: { $cond: [{ $eq: ["$status", "in-progress"] }, 1, 0] },
+              },
+              resolved: {
+                $sum: { $cond: [{ $eq: ["$status", "resolved"] }, 1, 0] },
+              },
+              closed: {
+                $sum: { $cond: [{ $eq: ["$status", "closed"] }, 1, 0] },
+              },
+              spam: { $sum: { $cond: [{ $eq: ["$status", "spam"] }, 1, 0] } },
             },
           },
         ],
@@ -976,7 +1046,7 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
         byCategory: [
           {
             $group: {
-              _id: '$category',
+              _id: "$category",
               count: { $sum: 1 },
             },
           },
@@ -986,7 +1056,7 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
         byPriority: [
           {
             $group: {
-              _id: '$priority',
+              _id: "$priority",
               count: { $sum: 1 },
             },
           },
@@ -996,7 +1066,11 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
           {
             $group: {
               _id: {
-                $cond: [{ $eq: ['$assignedTo', null] }, 'unassigned', 'assigned'],
+                $cond: [
+                  { $eq: ["$assignedTo", null] },
+                  "unassigned",
+                  "assigned",
+                ],
               },
               count: { $sum: 1 },
             },
@@ -1006,14 +1080,14 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
         responseTime: [
           {
             $match: {
-              status: { $in: ['resolved', 'closed'] },
-              'responses.0': { $exists: true },
+              status: { $in: ["resolved", "closed"] },
+              "responses.0": { $exists: true },
             },
           },
           {
             $project: {
               createdAt: 1,
-              firstResponse: { $arrayElemAt: ['$responses', 0] },
+              firstResponse: { $arrayElemAt: ["$responses", 0] },
               category: 1,
               priority: 1,
             },
@@ -1022,7 +1096,7 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
             $project: {
               responseHours: {
                 $divide: [
-                  { $subtract: ['$firstResponse.sentAt', '$createdAt'] },
+                  { $subtract: ["$firstResponse.sentAt", "$createdAt"] },
                   1000 * 60 * 60,
                 ],
               },
@@ -1033,19 +1107,19 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
           {
             $group: {
               _id: null,
-              avgHours: { $avg: '$responseHours' },
-              minHours: { $min: '$responseHours' },
-              maxHours: { $max: '$responseHours' },
+              avgHours: { $avg: "$responseHours" },
+              minHours: { $min: "$responseHours" },
+              maxHours: { $max: "$responseHours" },
               byCategory: {
                 $push: {
-                  category: '$category',
-                  hours: '$responseHours',
+                  category: "$category",
+                  hours: "$responseHours",
                 },
               },
               byPriority: {
                 $push: {
-                  priority: '$priority',
-                  hours: '$responseHours',
+                  priority: "$priority",
+                  hours: "$responseHours",
                 },
               },
             },
@@ -1054,11 +1128,11 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
         // Top responders
         topResponders: [
           {
-            $unwind: '$responses',
+            $unwind: "$responses",
           },
           {
             $group: {
-              _id: '$responses.responder',
+              _id: "$responses.responder",
               responseCount: { $sum: 1 },
             },
           },
@@ -1066,20 +1140,20 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
           { $limit: 10 },
           {
             $lookup: {
-              from: 'users',
-              localField: '_id',
-              foreignField: '_id',
-              as: 'userDetails',
+              from: "users",
+              localField: "_id",
+              foreignField: "_id",
+              as: "userDetails",
             },
           },
           {
-            $unwind: '$userDetails',
+            $unwind: "$userDetails",
           },
           {
             $project: {
-              responderId: '$_id',
-              responderName: '$userDetails.name',
-              responderRole: '$userDetails.role',
+              responderId: "$_id",
+              responderName: "$userDetails.name",
+              responderRole: "$userDetails.role",
               responseCount: 1,
             },
           },
@@ -1096,7 +1170,8 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
     byCategory: stats[0].byCategory,
     byPriority: stats[0].byPriority,
     byAssignment: stats[0].byAssignment,
-    responseTime: stats[0].responseTime.length > 0 ? stats[0].responseTime[0] : null,
+    responseTime:
+      stats[0].responseTime.length > 0 ? stats[0].responseTime[0] : null,
     topResponders: stats[0].topResponders,
   };
 
@@ -1105,3 +1180,22 @@ exports.getContactStats = asyncHandler(async (req, res, next) => {
     data: result,
   });
 });
+
+// Create controller object with all methods
+const contactController = {
+  submitContact,
+  getAllContacts,
+  getContact,
+  updateContactStatus,
+  assignContact,
+  respondToContact,
+  setFollowUp,
+  markAsSpam,
+  archiveContact,
+  deleteContact,
+  getContactsNeedingFollowUp,
+  getContactStats,
+};
+
+// Export the controller object as default
+export default contactController;
