@@ -1,21 +1,21 @@
 // server/src/controllers/profileController.js
-const User = require('../models/User');
-const DonationRequest = require('../models/DonationRequest');
-const Funding = require('../models/Funding');
-const ActivityLog = require('../models/ActivityLog');
-const Notification = require('../models/Notification');
-const asyncHandler = require('../middleware/asyncHandler');
-const ErrorResponse = require('../utils/errorResponse');
-const imageBB = require('../utils/imageBB');
+import User from "../models/User.js";
+import DonationRequest from "../models/DonationRequest.js";
+import Funding from "../models/Funding.js";
+import ActivityLog from "../models/ActivityLog.js";
+import Notification from "../models/Notification.js";
+import asyncHandler from "../middleware/asyncHandler.js";
+import ErrorResponse from "../utils/errorResponse.js";
+import imageBB from "../utils/imageBB.js";
 
 // @desc    Get user profile
 // @route   GET /api/profile
 // @access  Private
-exports.getProfile = asyncHandler(async (req, res, next) => {
-  const user = await User.findById(req.user.id).select('-password');
+export const getProfile = asyncHandler(async (req, res, next) => {
+  const user = await User.findById(req.user.id).select("-password");
 
   if (!user) {
-    return next(new ErrorResponse('User not found', 404));
+    return next(new ErrorResponse("User not found", 404));
   }
 
   // Get recent activity stats
@@ -23,15 +23,15 @@ exports.getProfile = asyncHandler(async (req, res, next) => {
     DonationRequest.find({ requester: user._id, isActive: true })
       .sort({ createdAt: -1 })
       .limit(3)
-      .select('recipientName bloodGroup donationDate status'),
+      .select("recipientName bloodGroup donationDate status"),
     DonationRequest.find({ donor: user._id, isActive: true })
       .sort({ donationDate: -1 })
       .limit(3)
-      .select('recipientName bloodGroup donationDate status'),
-    Funding.find({ donor: user._id, status: 'succeeded' })
+      .select("recipientName bloodGroup donationDate status"),
+    Funding.find({ donor: user._id, status: "succeeded" })
       .sort({ transactionDate: -1 })
       .limit(3)
-      .select('amount currency transactionDate'),
+      .select("amount currency transactionDate"),
   ]);
 
   const profileData = {
@@ -49,16 +49,16 @@ exports.getProfile = asyncHandler(async (req, res, next) => {
     userName: user.name,
     userEmail: user.email,
     userRole: user.role,
-    action: 'Viewed Profile',
-    actionType: 'read',
-    category: 'profile',
-    entityType: 'user',
+    action: "Viewed Profile",
+    actionType: "read",
+    category: "profile",
+    entityType: "user",
     entityId: user._id,
     entityName: user.name,
     description: `Viewed own profile`,
-    status: 'success',
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   res.status(200).json({
@@ -70,9 +70,9 @@ exports.getProfile = asyncHandler(async (req, res, next) => {
 // @desc    Update user profile
 // @route   PUT /api/profile
 // @access  Private
-exports.updateProfile = asyncHandler(async (req, res, next) => {
+export const updateProfile = asyncHandler(async (req, res, next) => {
   const updates = { ...req.body };
-  
+
   // Remove fields that cannot be updated via profile
   delete updates.email;
   delete updates.password;
@@ -83,17 +83,17 @@ exports.updateProfile = asyncHandler(async (req, res, next) => {
   if (updates.bloodGroup && updates.bloodGroup !== req.user.bloodGroup) {
     // Blood group changes require special handling (notifications, etc.)
     updates.bloodGroup = updates.bloodGroup.toUpperCase();
-    
+
     // Create notification about blood group change
     await Notification.createSystemNotification({
       recipient: req.user._id,
       recipientEmail: req.user.email,
-      title: 'Blood Group Updated',
+      title: "Blood Group Updated",
       message: `Your blood group has been changed from ${req.user.bloodGroup} to ${updates.bloodGroup}. This may affect donation compatibility.`,
-      type: 'info',
-      category: 'profile',
-      priority: 'medium',
-      actionUrl: '/dashboard/profile',
+      type: "info",
+      category: "profile",
+      priority: "medium",
+      actionUrl: "/dashboard/profile",
       data: {
         oldBloodGroup: req.user.bloodGroup,
         newBloodGroup: updates.bloodGroup,
@@ -107,21 +107,21 @@ exports.updateProfile = asyncHandler(async (req, res, next) => {
       const avatar = req.files.avatar;
       const uploadResult = await imageBB.uploadImage(avatar);
       updates.avatar = uploadResult.url;
-      
+
       // Create notification about avatar change
       await Notification.createSystemNotification({
         recipient: req.user._id,
         recipientEmail: req.user.email,
-        title: 'Profile Picture Updated',
-        message: 'Your profile picture has been successfully updated.',
-        type: 'success',
-        category: 'profile',
-        priority: 'low',
-        actionUrl: '/dashboard/profile',
+        title: "Profile Picture Updated",
+        message: "Your profile picture has been successfully updated.",
+        type: "success",
+        category: "profile",
+        priority: "low",
+        actionUrl: "/dashboard/profile",
         data: { avatarUpdated: true },
       });
     } catch (uploadError) {
-      console.error('Avatar upload error:', uploadError);
+      console.error("Avatar upload error:", uploadError);
       // Don't fail the request if avatar upload fails
     }
   }
@@ -139,11 +139,11 @@ exports.updateProfile = asyncHandler(async (req, res, next) => {
   const user = await User.findByIdAndUpdate(req.user.id, updates, {
     new: true,
     runValidators: true,
-  }).select('-password');
+  }).select("-password");
 
   // Determine what changed
   const changedFields = [];
-  Object.keys(updates).forEach(key => {
+  Object.keys(updates).forEach((key) => {
     if (oldUserData[key] !== user[key]) {
       changedFields.push(key);
     }
@@ -155,14 +155,14 @@ exports.updateProfile = asyncHandler(async (req, res, next) => {
     userName: user.name,
     userEmail: user.email,
     userRole: user.role,
-    action: 'Updated Profile',
-    actionType: 'update',
-    category: 'profile',
-    entityType: 'user',
+    action: "Updated Profile",
+    actionType: "update",
+    category: "profile",
+    entityType: "user",
     entityId: user._id,
     entityName: user.name,
     description: `Updated profile information`,
-    details: `Changed fields: ${changedFields.join(', ')}`,
+    details: `Changed fields: ${changedFields.join(", ")}`,
     changes: {
       before: oldUserData,
       after: {
@@ -174,39 +174,41 @@ exports.updateProfile = asyncHandler(async (req, res, next) => {
         avatar: user.avatar,
       },
     },
-    status: 'success',
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   res.status(200).json({
     success: true,
     data: user,
-    message: 'Profile updated successfully',
+    message: "Profile updated successfully",
   });
 });
 
 // @desc    Change password
 // @route   PUT /api/profile/password
 // @access  Private
-exports.changePassword = asyncHandler(async (req, res, next) => {
+export const changePassword = asyncHandler(async (req, res, next) => {
   const { currentPassword, newPassword, confirmPassword } = req.body;
 
   // Validate input
   if (!currentPassword || !newPassword || !confirmPassword) {
-    return next(new ErrorResponse('All password fields are required', 400));
+    return next(new ErrorResponse("All password fields are required", 400));
   }
 
   if (newPassword !== confirmPassword) {
-    return next(new ErrorResponse('New passwords do not match', 400));
+    return next(new ErrorResponse("New passwords do not match", 400));
   }
 
   if (newPassword.length < 6) {
-    return next(new ErrorResponse('Password must be at least 6 characters', 400));
+    return next(
+      new ErrorResponse("Password must be at least 6 characters", 400)
+    );
   }
 
   // Get user with password
-  const user = await User.findById(req.user.id).select('+password');
+  const user = await User.findById(req.user.id).select("+password");
 
   // Check current password
   const isMatch = await user.comparePassword(currentPassword);
@@ -218,20 +220,21 @@ exports.changePassword = asyncHandler(async (req, res, next) => {
       userName: user.name,
       userEmail: user.email,
       userRole: user.role,
-      action: 'Failed Password Change Attempt',
-      actionType: 'security',
-      category: 'security',
-      entityType: 'user',
+      action: "Failed Password Change Attempt",
+      actionType: "security",
+      category: "security",
+      entityType: "user",
       entityId: user._id,
       entityName: user.name,
-      description: 'Failed password change attempt - incorrect current password',
-      status: 'failed',
-      severity: 'warning',
+      description:
+        "Failed password change attempt - incorrect current password",
+      status: "failed",
+      severity: "warning",
       userIp: req.ip,
-      userAgent: req.headers['user-agent'],
+      userAgent: req.headers["user-agent"],
     });
 
-    return next(new ErrorResponse('Current password is incorrect', 401));
+    return next(new ErrorResponse("Current password is incorrect", 401));
   }
 
   // Update password
@@ -244,46 +247,48 @@ exports.changePassword = asyncHandler(async (req, res, next) => {
     userName: user.name,
     userEmail: user.email,
     userRole: user.role,
-    action: 'Password Changed',
-    actionType: 'update',
-    category: 'security',
-    entityType: 'user',
+    action: "Password Changed",
+    actionType: "update",
+    category: "security",
+    entityType: "user",
     entityId: user._id,
     entityName: user.name,
-    description: 'Successfully changed password',
-    status: 'success',
-    severity: 'info',
+    description: "Successfully changed password",
+    status: "success",
+    severity: "info",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   // Create notification
   await Notification.createSystemNotification({
     recipient: user._id,
     recipientEmail: user.email,
-    title: 'Password Changed Successfully 🔒',
+    title: "Password Changed Successfully 🔒",
     message: `Your password was changed successfully on ${new Date().toLocaleDateString()}. If you didn't make this change, please contact support immediately.`,
-    type: 'info',
-    category: 'security',
-    priority: 'medium',
-    actionUrl: '/dashboard/profile',
+    type: "info",
+    category: "security",
+    priority: "medium",
+    actionUrl: "/dashboard/profile",
     data: { passwordChanged: true, timestamp: new Date() },
   });
 
   res.status(200).json({
     success: true,
-    message: 'Password changed successfully',
+    message: "Password changed successfully",
   });
 });
 
 // @desc    Update location
 // @route   PUT /api/profile/location
 // @access  Private
-exports.updateLocation = asyncHandler(async (req, res, next) => {
+export const updateLocation = asyncHandler(async (req, res, next) => {
   const { district, upazila } = req.body;
 
   if (!district || !upazila) {
-    return next(new ErrorResponse('Both district and upazila are required', 400));
+    return next(
+      new ErrorResponse("Both district and upazila are required", 400)
+    );
   }
 
   const oldLocation = {
@@ -298,7 +303,7 @@ exports.updateLocation = asyncHandler(async (req, res, next) => {
       new: true,
       runValidators: true,
     }
-  ).select('-password');
+  ).select("-password");
 
   // Log activity
   await ActivityLog.logActivity({
@@ -306,10 +311,10 @@ exports.updateLocation = asyncHandler(async (req, res, next) => {
     userName: user.name,
     userEmail: user.email,
     userRole: user.role,
-    action: 'Updated Location',
-    actionType: 'update',
-    category: 'profile',
-    entityType: 'user',
+    action: "Updated Location",
+    actionType: "update",
+    category: "profile",
+    entityType: "user",
     entityId: user._id,
     entityName: user.name,
     description: `Updated location from ${oldLocation.upazila}, ${oldLocation.district} to ${upazila}, ${district}`,
@@ -317,35 +322,35 @@ exports.updateLocation = asyncHandler(async (req, res, next) => {
       before: oldLocation,
       after: { district, upazila },
     },
-    status: 'success',
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   // Create notification
   await Notification.createSystemNotification({
     recipient: user._id,
     recipientEmail: user.email,
-    title: 'Location Updated 📍',
+    title: "Location Updated 📍",
     message: `Your location has been updated to ${upazila}, ${district}. This helps us match you with nearby donation requests.`,
-    type: 'info',
-    category: 'profile',
-    priority: 'low',
-    actionUrl: '/dashboard/profile',
+    type: "info",
+    category: "profile",
+    priority: "low",
+    actionUrl: "/dashboard/profile",
     data: { locationUpdated: true },
   });
 
   res.status(200).json({
     success: true,
     data: user,
-    message: 'Location updated successfully',
+    message: "Location updated successfully",
   });
 });
 
 // @desc    Toggle availability status
 // @route   PATCH /api/profile/availability
 // @access  Private
-exports.toggleAvailability = asyncHandler(async (req, res, next) => {
+export const toggleAvailability = asyncHandler(async (req, res, next) => {
   const user = await User.findById(req.user.id);
 
   const newAvailability = !user.isAvailable;
@@ -358,44 +363,50 @@ exports.toggleAvailability = asyncHandler(async (req, res, next) => {
     userName: user.name,
     userEmail: user.email,
     userRole: user.role,
-    action: 'Updated Availability Status',
-    actionType: 'update',
-    category: 'profile',
-    entityType: 'user',
+    action: "Updated Availability Status",
+    actionType: "update",
+    category: "profile",
+    entityType: "user",
     entityId: user._id,
     entityName: user.name,
-    description: `Set availability to ${newAvailability ? 'available' : 'unavailable'}`,
-    status: 'success',
+    description: `Set availability to ${
+      newAvailability ? "available" : "unavailable"
+    }`,
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   // Create notification
   await Notification.createSystemNotification({
     recipient: user._id,
     recipientEmail: user.email,
-    title: newAvailability ? 'You Are Now Available for Donations! 🩸' : 'You Are Now Unavailable',
-    message: newAvailability 
-      ? 'You have marked yourself as available for blood donations. You will now receive notifications for nearby donation requests.'
-      : 'You have marked yourself as unavailable for blood donations. You will not receive donation request notifications.',
-    type: newAvailability ? 'success' : 'warning',
-    category: 'profile',
-    priority: 'medium',
-    actionUrl: '/dashboard/profile',
+    title: newAvailability
+      ? "You Are Now Available for Donations! 🩸"
+      : "You Are Now Unavailable",
+    message: newAvailability
+      ? "You have marked yourself as available for blood donations. You will now receive notifications for nearby donation requests."
+      : "You have marked yourself as unavailable for blood donations. You will not receive donation request notifications.",
+    type: newAvailability ? "success" : "warning",
+    category: "profile",
+    priority: "medium",
+    actionUrl: "/dashboard/profile",
     data: { isAvailable: newAvailability },
   });
 
   res.status(200).json({
     success: true,
     data: { isAvailable: newAvailability },
-    message: `You are now ${newAvailability ? 'available' : 'unavailable'} for donations`,
+    message: `You are now ${
+      newAvailability ? "available" : "unavailable"
+    } for donations`,
   });
 });
 
 // @desc    Get profile statistics
 // @route   GET /api/profile/stats
 // @access  Private
-exports.getProfileStats = asyncHandler(async (req, res, next) => {
+export const getProfileStats = asyncHandler(async (req, res, next) => {
   const user = req.user;
 
   // Get comprehensive statistics
@@ -416,7 +427,7 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
       },
       {
         $group: {
-          _id: '$status',
+          _id: "$status",
           count: { $sum: 1 },
         },
       },
@@ -432,7 +443,7 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
       },
       {
         $group: {
-          _id: '$status',
+          _id: "$status",
           count: { $sum: 1 },
         },
       },
@@ -443,13 +454,13 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
       {
         $match: {
           donor: user._id,
-          status: 'succeeded',
+          status: "succeeded",
         },
       },
       {
         $group: {
           _id: null,
-          totalAmount: { $sum: '$amount' },
+          totalAmount: { $sum: "$amount" },
           count: { $sum: 1 },
         },
       },
@@ -458,14 +469,14 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
     // Blood compatibility (could be enhanced with actual medical data)
     (() => {
       const compatibilityChart = {
-        'A+': ['A+', 'AB+'],
-        'A-': ['A+', 'A-', 'AB+', 'AB-'],
-        'B+': ['B+', 'AB+'],
-        'B-': ['B+', 'B-', 'AB+', 'AB-'],
-        'AB+': ['AB+'],
-        'AB-': ['AB+', 'AB-'],
-        'O+': ['A+', 'B+', 'AB+', 'O+'],
-        'O-': ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-'],
+        "A+": ["A+", "AB+"],
+        "A-": ["A+", "A-", "AB+", "AB-"],
+        "B+": ["B+", "AB+"],
+        "B-": ["B+", "B-", "AB+", "AB-"],
+        "AB+": ["AB+"],
+        "AB-": ["AB+", "AB-"],
+        "O+": ["A+", "B+", "AB+", "O+"],
+        "O-": ["A+", "A-", "B+", "B-", "AB+", "AB-", "O+", "O-"],
       };
       return compatibilityChart[user.bloodGroup] || [];
     })(),
@@ -477,16 +488,18 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
           canDonate: true,
           daysSinceLastDonation: null,
           nextEligibleDate: null,
-          message: 'You have never donated before. You are eligible to donate.',
+          message: "You have never donated before. You are eligible to donate.",
         };
       }
 
       const lastDonation = new Date(user.lastDonationDate);
       const now = new Date();
-      const daysSinceLastDonation = Math.floor((now - lastDonation) / (1000 * 60 * 60 * 24));
+      const daysSinceLastDonation = Math.floor(
+        (now - lastDonation) / (1000 * 60 * 60 * 24)
+      );
       const daysRequired = 90; // Minimum days between donations
       const canDonate = daysSinceLastDonation >= daysRequired;
-      
+
       const nextEligibleDate = new Date(lastDonation);
       nextEligibleDate.setDate(nextEligibleDate.getDate() + daysRequired);
 
@@ -494,7 +507,7 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
         canDonate,
         daysSinceLastDonation,
         nextEligibleDate,
-        message: canDonate 
+        message: canDonate
           ? `You are eligible to donate. It has been ${daysSinceLastDonation} days since your last donation.`
           : `You can donate again after ${nextEligibleDate.toLocaleDateString()}. It has been ${daysSinceLastDonation} days since your last donation.`,
       };
@@ -510,7 +523,7 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
     canceled: 0,
   };
 
-  donationRequests.forEach(stat => {
+  donationRequests.forEach((stat) => {
     requestStats[stat._id] = stat.count;
     requestStats.total += stat.count;
   });
@@ -522,13 +535,14 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
     canceled: 0,
   };
 
-  donationsMade.forEach(stat => {
+  donationsMade.forEach((stat) => {
     donationStats[stat._id] = stat.count;
     donationStats.total += stat.count;
   });
 
   // Format funding stats
-  const fundingResult = fundingStats.length > 0 ? fundingStats[0] : { totalAmount: 0, count: 0 };
+  const fundingResult =
+    fundingStats.length > 0 ? fundingStats[0] : { totalAmount: 0, count: 0 };
 
   const stats = {
     user: {
@@ -548,12 +562,12 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
     funding: {
       totalDonations: fundingResult.count,
       totalAmount: fundingResult.totalAmount,
-      formattedAmount: `৳${fundingResult.totalAmount.toLocaleString('en-BD')}`,
+      formattedAmount: `৳${fundingResult.totalAmount.toLocaleString("en-BD")}`,
     },
     compatibility: {
       canDonateTo: bloodCompatibility,
-      canReceiveFrom: ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-'].filter(group => 
-        bloodCompatibility.includes(group)
+      canReceiveFrom: ["A+", "A-", "B+", "B-", "AB+", "AB-", "O+", "O-"].filter(
+        (group) => bloodCompatibility.includes(group)
       ),
     },
     eligibility: await eligibility,
@@ -576,7 +590,7 @@ exports.getProfileStats = asyncHandler(async (req, res, next) => {
 // @desc    Get donation history
 // @route   GET /api/profile/donation-history
 // @access  Private
-exports.getDonationHistory = asyncHandler(async (req, res, next) => {
+export const getDonationHistory = asyncHandler(async (req, res, next) => {
   const page = parseInt(req.query.page, 10) || 1;
   const limit = parseInt(req.query.limit, 10) || 10;
   const skip = (page - 1) * limit;
@@ -590,8 +604,10 @@ exports.getDonationHistory = asyncHandler(async (req, res, next) => {
 
   const [donations, total] = await Promise.all([
     DonationRequest.find(filter)
-      .populate('requester', 'name email avatar phone')
-      .select('recipientName bloodGroup hospitalName donationDate donationTime status')
+      .populate("requester", "name email avatar phone")
+      .select(
+        "recipientName bloodGroup hospitalName donationDate donationTime status"
+      )
       .skip(skip)
       .limit(limit)
       .sort({ donationDate: -1 }),
@@ -599,7 +615,7 @@ exports.getDonationHistory = asyncHandler(async (req, res, next) => {
   ]);
 
   // Calculate life impact
-  const lifeImpact = donations.filter(d => d.status === 'done').length;
+  const lifeImpact = donations.filter((d) => d.status === "done").length;
   const estimatedLivesSaved = lifeImpact * 3; // Each donation can save up to 3 lives
 
   res.status(200).json({
@@ -625,7 +641,7 @@ exports.getDonationHistory = asyncHandler(async (req, res, next) => {
 // @desc    Get request history
 // @route   GET /api/profile/request-history
 // @access  Private
-exports.getRequestHistory = asyncHandler(async (req, res, next) => {
+export const getRequestHistory = asyncHandler(async (req, res, next) => {
   const page = parseInt(req.query.page, 10) || 1;
   const limit = parseInt(req.query.limit, 10) || 10;
   const skip = (page - 1) * limit;
@@ -639,8 +655,10 @@ exports.getRequestHistory = asyncHandler(async (req, res, next) => {
 
   const [requests, total] = await Promise.all([
     DonationRequest.find(filter)
-      .populate('donor', 'name email avatar phone bloodGroup')
-      .select('recipientName bloodGroup hospitalName donationDate donationTime status donor')
+      .populate("donor", "name email avatar phone bloodGroup")
+      .select(
+        "recipientName bloodGroup hospitalName donationDate donationTime status donor"
+      )
       .skip(skip)
       .limit(limit)
       .sort({ donationDate: -1 }),
@@ -649,8 +667,11 @@ exports.getRequestHistory = asyncHandler(async (req, res, next) => {
 
   // Calculate success rate
   const totalRequests = requests.length;
-  const completedRequests = requests.filter(r => r.status === 'done').length;
-  const successRate = totalRequests > 0 ? Math.round((completedRequests / totalRequests) * 100) : 0;
+  const completedRequests = requests.filter((r) => r.status === "done").length;
+  const successRate =
+    totalRequests > 0
+      ? Math.round((completedRequests / totalRequests) * 100)
+      : 0;
 
   res.status(200).json({
     success: true,
@@ -660,7 +681,7 @@ exports.getRequestHistory = asyncHandler(async (req, res, next) => {
       totalRequests,
       completedRequests,
       successRate: `${successRate}%`,
-      pendingRequests: requests.filter(r => r.status === 'pending').length,
+      pendingRequests: requests.filter((r) => r.status === "pending").length,
     },
     pagination: {
       page,
@@ -676,7 +697,7 @@ exports.getRequestHistory = asyncHandler(async (req, res, next) => {
 // @desc    Get funding history
 // @route   GET /api/profile/funding-history
 // @access  Private
-exports.getFundingHistory = asyncHandler(async (req, res, next) => {
+export const getFundingHistory = asyncHandler(async (req, res, next) => {
   const page = parseInt(req.query.page, 10) || 1;
   const limit = parseInt(req.query.limit, 10) || 10;
   const skip = (page - 1) * limit;
@@ -689,7 +710,7 @@ exports.getFundingHistory = asyncHandler(async (req, res, next) => {
 
   const [fundings, total] = await Promise.all([
     Funding.find(filter)
-      .select('amount currency transactionDate status receiptUrl message')
+      .select("amount currency transactionDate status receiptUrl message")
       .skip(skip)
       .limit(limit)
       .sort({ transactionDate: -1 }),
@@ -699,7 +720,7 @@ exports.getFundingHistory = asyncHandler(async (req, res, next) => {
   // Calculate totals
   const totals = fundings.reduce(
     (acc, funding) => {
-      if (funding.status === 'succeeded') {
+      if (funding.status === "succeeded") {
         acc.successful += funding.amount;
         acc.count += 1;
       }
@@ -714,7 +735,7 @@ exports.getFundingHistory = asyncHandler(async (req, res, next) => {
     total,
     totals: {
       amount: totals.successful,
-      formattedAmount: `৳${totals.successful.toLocaleString('en-BD')}`,
+      formattedAmount: `৳${totals.successful.toLocaleString("en-BD")}`,
       donations: totals.count,
     },
     pagination: {
@@ -731,74 +752,80 @@ exports.getFundingHistory = asyncHandler(async (req, res, next) => {
 // @desc    Update notification preferences
 // @route   PUT /api/profile/notifications
 // @access  Private
-exports.updateNotificationPreferences = asyncHandler(async (req, res, next) => {
-  const { preferences } = req.body;
+export const updateNotificationPreferences = asyncHandler(
+  async (req, res, next) => {
+    const { preferences } = req.body;
 
-  if (!preferences || typeof preferences !== 'object') {
-    return next(new ErrorResponse('Notification preferences are required', 400));
+    if (!preferences || typeof preferences !== "object") {
+      return next(
+        new ErrorResponse("Notification preferences are required", 400)
+      );
+    }
+
+    const user = await User.findByIdAndUpdate(
+      req.user.id,
+      { notificationPreferences: preferences },
+      { new: true }
+    ).select("-password");
+
+    // Log activity
+    await ActivityLog.logActivity({
+      user: user._id,
+      userName: user.name,
+      userEmail: user.email,
+      userRole: user.role,
+      action: "Updated Notification Preferences",
+      actionType: "update",
+      category: "profile",
+      entityType: "user",
+      entityId: user._id,
+      entityName: user.name,
+      description: "Updated notification preferences",
+      details: JSON.stringify(preferences),
+      status: "success",
+      userIp: req.ip,
+      userAgent: req.headers["user-agent"],
+    });
+
+    res.status(200).json({
+      success: true,
+      data: { notificationPreferences: user.notificationPreferences },
+      message: "Notification preferences updated successfully",
+    });
   }
-
-  const user = await User.findByIdAndUpdate(
-    req.user.id,
-    { notificationPreferences: preferences },
-    { new: true }
-  ).select('-password');
-
-  // Log activity
-  await ActivityLog.logActivity({
-    user: user._id,
-    userName: user.name,
-    userEmail: user.email,
-    userRole: user.role,
-    action: 'Updated Notification Preferences',
-    actionType: 'update',
-    category: 'profile',
-    entityType: 'user',
-    entityId: user._id,
-    entityName: user.name,
-    description: 'Updated notification preferences',
-    details: JSON.stringify(preferences),
-    status: 'success',
-    userIp: req.ip,
-    userAgent: req.headers['user-agent'],
-  });
-
-  res.status(200).json({
-    success: true,
-    data: { notificationPreferences: user.notificationPreferences },
-    message: 'Notification preferences updated successfully',
-  });
-});
+);
 
 // @desc    Deactivate account (soft delete)
 // @route   DELETE /api/profile/deactivate
 // @access  Private
-exports.deactivateAccount = asyncHandler(async (req, res, next) => {
+export const deactivateAccount = asyncHandler(async (req, res, next) => {
   const user = await User.findById(req.user.id);
 
   // Check if user has active donations or requests
   const [activeDonations, activeRequests] = await Promise.all([
     DonationRequest.countDocuments({
       donor: user._id,
-      status: { $in: ['pending', 'inprogress'] },
+      status: { $in: ["pending", "inprogress"] },
       isActive: true,
     }),
     DonationRequest.countDocuments({
       requester: user._id,
-      status: { $in: ['pending', 'inprogress'] },
+      status: { $in: ["pending", "inprogress"] },
       isActive: true,
     }),
   ]);
 
   if (activeDonations > 0 || activeRequests > 0) {
-    return next(new ErrorResponse(
-      'Cannot deactivate account with active donations or requests. Please complete or cancel them first.',
-      400
-    ));
+    return next(
+      new ErrorResponse(
+        "Cannot deactivate account with active donations or requests. Please complete or cancel them first.",
+        400
+      )
+    );
   }
 
   // Soft delete - mark as inactive
-  user.status = 'inactive';
+  user.status = "inactive";
   user.isAvailable = false;
   user.deactivatedAt = new Date();
   await user.save();
@@ -806,12 +833,28 @@ exports.deactivateAccount = asyncHandler(async (req, res, next) => {
   // Log all active donations and requests as canceled
   await Promise.all([
     DonationRequest.updateMany(
-      { donor: user._id, status: 'pending' },
-      { status: 'canceled', $push: { statusHistory: { status: 'canceled', note: 'Donor deactivated account' } } }
+      { donor: user._id, status: "pending" },
+      {
+        status: "canceled",
+        $push: {
+          statusHistory: {
+            status: "canceled",
+            note: "Donor deactivated account",
+          },
+        },
+      }
     ),
     DonationRequest.updateMany(
-      { requester: user._id, status: 'pending' },
-      { status: 'canceled', $push: { statusHistory: { status: 'canceled', note: 'Requester deactivated account' } } }
+      { requester: user._id, status: "pending" },
+      {
+        status: "canceled",
+        $push: {
+          statusHistory: {
+            status: "canceled",
+            note: "Requester deactivated account",
+          },
+        },
+      }
     ),
   ]);
 
@@ -821,27 +864,28 @@ exports.deactivateAccount = asyncHandler(async (req, res, next) => {
     userName: user.name,
     userEmail: user.email,
     userRole: user.role,
-    action: 'Deactivated Account',
-    actionType: 'delete',
-    category: 'profile',
-    entityType: 'user',
+    action: "Deactivated Account",
+    actionType: "delete",
+    category: "profile",
+    entityType: "user",
     entityId: user._id,
     entityName: user.name,
-    description: 'Deactivated user account',
-    status: 'success',
-    severity: 'warning',
+    description: "Deactivated user account",
+    status: "success",
+    severity: "warning",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
   // Clear token cookie
-  res.cookie('token', 'none', {
+  res.cookie("token", "none", {
     expires: new Date(Date.now() + 10 * 1000),
     httpOnly: true,
   });
 
   res.status(200).json({
     success: true,
-    message: 'Account deactivated successfully. You can reactivate by logging in within 30 days.',
+    message:
+      "Account deactivated successfully. You can reactivate by logging in within 30 days.",
   });
 });
