@@ -1,25 +1,24 @@
 // server/src/controllers/searchController.js
-const User = require('../models/User');
-const DonationRequest = require('../models/DonationRequest');
-const Funding = require('../models/Funding');
-const Contact = require('../models/Contact');
-const ActivityLog = require('../models/ActivityLog');
-const asyncHandler = require('../middleware/asyncHandler');
-const ErrorResponse = require('../utils/errorResponse');
-const { districts, upazilas } = require('../utils/locationData');
+import User from "../models/User.js";
+import DonationRequest from "../models/DonationRequest.js";
+import Funding from "../models/Funding.js";
+import Contact from "../models/Contact.js";
+import ActivityLog from "../models/ActivityLog.js";
+import asyncHandler from "../middleware/asyncHandler.js";
+import ErrorResponse from "../utils/errorResponse.js";
 
 // @desc    Search donors with filters
 // @route   GET /api/search/donors
 // @access  Public
-exports.searchDonors = asyncHandler(async (req, res, next) => {
+export const searchDonors = asyncHandler(async (req, res, next) => {
   const {
     bloodGroup,
     district,
     upazila,
     page = 1,
     limit = 20,
-    sortBy = 'relevance',
-    sortOrder = 'desc',
+    sortBy = "relevance",
+    sortOrder = "desc",
     availableOnly = true,
   } = req.query;
 
@@ -29,54 +28,54 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
 
   // Build filter for donors
   const filter = {
-    role: 'donor',
-    status: 'active',
+    role: "donor",
+    status: "active",
   };
 
   // Apply blood group filter
-  if (bloodGroup && bloodGroup !== 'all') {
+  if (bloodGroup && bloodGroup !== "all") {
     filter.bloodGroup = bloodGroup.toUpperCase();
   }
 
   // Apply location filters
-  if (district && district !== 'all') {
+  if (district && district !== "all") {
     filter.district = district;
   }
-  
-  if (upazila && upazila !== 'all') {
+
+  if (upazila && upazila !== "all") {
     filter.upazila = upazila;
   }
 
   // Apply availability filter
-  if (availableOnly === 'true') {
+  if (availableOnly === "true") {
     filter.isAvailable = true;
-    
+
     // Also check eligibility based on last donation date
     filter.$or = [
       { lastDonationDate: null },
       {
         lastDonationDate: {
-          $lt: new Date(Date.now() - 90 * 24 * 60 * 60 * 1000) // 90 days ago
-        }
-      }
+          $lt: new Date(Date.now() - 90 * 24 * 60 * 60 * 1000), // 90 days ago
+        },
+      },
     ];
   }
 
   // Sorting logic
   const sort = {};
   switch (sortBy) {
-    case 'recent':
-      sort.createdAt = sortOrder === 'desc' ? -1 : 1;
+    case "recent":
+      sort.createdAt = sortOrder === "desc" ? -1 : 1;
       break;
-    case 'donations':
-      sort.totalDonations = sortOrder === 'desc' ? -1 : 1;
+    case "donations":
+      sort.totalDonations = sortOrder === "desc" ? -1 : 1;
       break;
-    case 'name':
-      sort.name = sortOrder === 'desc' ? -1 : 1;
+    case "name":
+      sort.name = sortOrder === "desc" ? -1 : 1;
       break;
-    case 'location':
-      sort.district = sortOrder === 'desc' ? -1 : 1;
-      sort.upazila = sortOrder === 'desc' ? -1 : 1;
+    case "location":
+      sort.district = sortOrder === "desc" ? -1 : 1;
+      sort.upazila = sortOrder === "desc" ? -1 : 1;
       break;
     default: // relevance - prioritize available donors with matching blood group
       sort.isAvailable = -1;
@@ -87,7 +86,7 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
   // Execute query with pagination
   const [donors, total] = await Promise.all([
     User.find(filter)
-      .select('-password -notificationPreferences')
+      .select("-password -notificationPreferences")
       .skip(skip)
       .limit(limitNum)
       .sort(sort),
@@ -98,55 +97,57 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
   const enrichedDonors = await Promise.all(
     donors.map(async (donor) => {
       const donorObj = donor.toObject();
-      
+
       // Calculate eligibility
       let isEligible = true;
-      let eligibilityMessage = 'Available for donation';
+      let eligibilityMessage = "Available for donation";
       let daysSinceLastDonation = null;
-      
+
       if (donor.lastDonationDate) {
         daysSinceLastDonation = Math.floor(
-          (new Date() - new Date(donor.lastDonationDate)) / (1000 * 60 * 60 * 24)
+          (new Date() - new Date(donor.lastDonationDate)) /
+            (1000 * 60 * 60 * 24)
         );
-        
+
         if (daysSinceLastDonation < 90) {
           isEligible = false;
           const daysLeft = 90 - daysSinceLastDonation;
           eligibilityMessage = `Can donate in ${daysLeft} days`;
         }
       }
-      
+
       // Get donor's recent donation history
       const recentDonations = await DonationRequest.find({
         donor: donor._id,
-        status: 'done',
+        status: "done",
         isActive: true,
       })
         .sort({ donationDate: -1 })
         .limit(3)
-        .select('recipientName donationDate bloodGroup');
-      
+        .select("recipientName donationDate bloodGroup");
+
       // Get donor's location details
       const locationDetails = {
         district: donor.district,
         upazila: donor.upazila,
         fullAddress: `${donor.upazila}, ${donor.district}`,
       };
-      
+
       // Calculate response rate (if donor has been contacted before)
       const acceptedRequests = await DonationRequest.countDocuments({
         donor: donor._id,
-        status: 'done',
+        status: "done",
       });
-      
+
       const totalRequests = await DonationRequest.countDocuments({
         donor: donor._id,
       });
-      
-      const responseRate = totalRequests > 0 
-        ? Math.round((acceptedRequests / totalRequests) * 100) 
-        : 100;
-      
+
+      const responseRate =
+        totalRequests > 0
+          ? Math.round((acceptedRequests / totalRequests) * 100)
+          : 100;
+
       return {
         ...donorObj,
         eligibility: {
@@ -155,14 +156,17 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
           lastDonationDate: donor.lastDonationDate,
           daysSinceLastDonation,
           nextEligibleDate: donor.lastDonationDate
-            ? new Date(new Date(donor.lastDonationDate).getTime() + 90 * 24 * 60 * 60 * 1000)
+            ? new Date(
+                new Date(donor.lastDonationDate).getTime() +
+                  90 * 24 * 60 * 60 * 1000
+              )
             : null,
         },
         statistics: {
           totalDonations: donor.totalDonations || 0,
           recentDonations: recentDonations.length,
           responseRate: `${responseRate}%`,
-          successRate: donor.totalDonations > 0 ? '100%' : '0%',
+          successRate: donor.totalDonations > 0 ? "100%" : "0%",
         },
         location: locationDetails,
         recentActivity: recentDonations,
@@ -178,11 +182,13 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
   const searchStats = await User.aggregate([
     {
       $match: {
-        role: 'donor',
-        status: 'active',
-        ...(bloodGroup && bloodGroup !== 'all' ? { bloodGroup: bloodGroup.toUpperCase() } : {}),
-        ...(district && district !== 'all' ? { district } : {}),
-        ...(upazila && upazila !== 'all' ? { upazila } : {}),
+        role: "donor",
+        status: "active",
+        ...(bloodGroup && bloodGroup !== "all"
+          ? { bloodGroup: bloodGroup.toUpperCase() }
+          : {}),
+        ...(district && district !== "all" ? { district } : {}),
+        ...(upazila && upazila !== "all" ? { upazila } : {}),
       },
     },
     {
@@ -190,20 +196,30 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
         bloodGroupStats: [
           {
             $group: {
-              _id: '$bloodGroup',
+              _id: "$bloodGroup",
               count: { $sum: 1 },
               available: {
                 $sum: {
                   $cond: [
                     {
                       $and: [
-                        { $eq: ['$isAvailable', true] },
+                        { $eq: ["$isAvailable", true] },
                         {
                           $or: [
-                            { $eq: ['$lastDonationDate', null] },
+                            { $eq: ["$lastDonationDate", null] },
                             {
                               $lt: [
-                                { $divide: [{ $subtract: [new Date(), '$lastDonationDate'] }, 1000 * 60 * 60 * 24] },
+                                {
+                                  $divide: [
+                                    {
+                                      $subtract: [
+                                        new Date(),
+                                        "$lastDonationDate",
+                                      ],
+                                    },
+                                    1000 * 60 * 60 * 24,
+                                  ],
+                                },
                                 90,
                               ],
                             },
@@ -224,8 +240,8 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
           {
             $group: {
               _id: {
-                district: '$district',
-                upazila: '$upazila',
+                district: "$district",
+                upazila: "$upazila",
               },
               count: { $sum: 1 },
             },
@@ -243,13 +259,23 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
                   $cond: [
                     {
                       $and: [
-                        { $eq: ['$isAvailable', true] },
+                        { $eq: ["$isAvailable", true] },
                         {
                           $or: [
-                            { $eq: ['$lastDonationDate', null] },
+                            { $eq: ["$lastDonationDate", null] },
                             {
                               $lt: [
-                                { $divide: [{ $subtract: [new Date(), '$lastDonationDate'] }, 1000 * 60 * 60 * 24] },
+                                {
+                                  $divide: [
+                                    {
+                                      $subtract: [
+                                        new Date(),
+                                        "$lastDonationDate",
+                                      ],
+                                    },
+                                    1000 * 60 * 60 * 24,
+                                  ],
+                                },
                                 90,
                               ],
                             },
@@ -276,14 +302,16 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
       userName: req.user.name,
       userEmail: req.user.email,
       userRole: req.user.role,
-      action: 'Searched for Donors',
-      actionType: 'search',
-      category: 'search',
-      description: `Searched donors with filters: Blood Group=${bloodGroup || 'Any'}, District=${district || 'Any'}, Upazila=${upazila || 'Any'}`,
+      action: "Searched for Donors",
+      actionType: "search",
+      category: "search",
+      description: `Searched donors with filters: Blood Group=${
+        bloodGroup || "Any"
+      }, District=${district || "Any"}, Upazila=${upazila || "Any"}`,
       details: `Found ${total} donors matching criteria`,
-      status: 'success',
+      status: "success",
       userIp: req.ip,
-      userAgent: req.headers['user-agent'],
+      userAgent: req.headers["user-agent"],
       request: {
         method: req.method,
         url: req.originalUrl,
@@ -297,10 +325,10 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
     count: enrichedDonors.length,
     total,
     filters: {
-      bloodGroup: bloodGroup || 'Any',
-      district: district || 'Any',
-      upazila: upazila || 'Any',
-      availableOnly: availableOnly === 'true',
+      bloodGroup: bloodGroup || "Any",
+      district: district || "Any",
+      upazila: upazila || "Any",
+      availableOnly: availableOnly === "true",
       sortBy,
       sortOrder,
     },
@@ -319,17 +347,17 @@ exports.searchDonors = asyncHandler(async (req, res, next) => {
 // @desc    Search donation requests
 // @route   GET /api/search/donation-requests
 // @access  Public
-exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
+export const searchDonationRequests = asyncHandler(async (req, res, next) => {
   const {
     bloodGroup,
     district,
     upazila,
-    status = 'pending',
+    status = "pending",
     urgency,
     page = 1,
     limit = 20,
-    sortBy = 'urgency',
-    sortOrder = 'desc',
+    sortBy = "urgency",
+    sortOrder = "desc",
   } = req.query;
 
   const pageNum = parseInt(page, 10);
@@ -343,21 +371,21 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
   };
 
   // Apply blood group filter
-  if (bloodGroup && bloodGroup !== 'all') {
+  if (bloodGroup && bloodGroup !== "all") {
     filter.bloodGroup = bloodGroup.toUpperCase();
   }
 
   // Apply location filters
-  if (district && district !== 'all') {
+  if (district && district !== "all") {
     filter.recipientDistrict = district;
   }
-  
-  if (upazila && upazila !== 'all') {
+
+  if (upazila && upazila !== "all") {
     filter.recipientUpazila = upazila;
   }
 
   // Apply urgency filter
-  if (urgency && urgency !== 'all') {
+  if (urgency && urgency !== "all") {
     filter.urgency = urgency;
   }
 
@@ -371,25 +399,25 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
   // Sorting logic
   const sort = {};
   switch (sortBy) {
-    case 'date':
-      sort.donationDate = sortOrder === 'desc' ? -1 : 1;
+    case "date":
+      sort.donationDate = sortOrder === "desc" ? -1 : 1;
       break;
-    case 'created':
-      sort.createdAt = sortOrder === 'desc' ? -1 : 1;
+    case "created":
+      sort.createdAt = sortOrder === "desc" ? -1 : 1;
       break;
-    case 'hospital':
-      sort.hospitalName = sortOrder === 'desc' ? -1 : 1;
+    case "hospital":
+      sort.hospitalName = sortOrder === "desc" ? -1 : 1;
       break;
     default: // urgency
-      sort.urgency = sortOrder === 'desc' ? -1 : 1;
+      sort.urgency = sortOrder === "desc" ? -1 : 1;
       sort.donationDate = 1; // Soonest first
   }
 
   // Execute query
   const [requests, total] = await Promise.all([
     DonationRequest.find(filter)
-      .populate('requester', 'name email avatar phone')
-      .populate('donor', 'name email avatar')
+      .populate("requester", "name email avatar phone")
+      .populate("donor", "name email avatar")
       .skip(skip)
       .limit(limitNum)
       .sort(sort),
@@ -402,10 +430,14 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
       $match: {
         isActive: true,
         status: status,
-        ...(bloodGroup && bloodGroup !== 'all' ? { bloodGroup: bloodGroup.toUpperCase() } : {}),
-        ...(district && district !== 'all' ? { recipientDistrict: district } : {}),
-        ...(upazila && upazila !== 'all' ? { recipientUpazila: upazila } : {}),
-        ...(urgency && urgency !== 'all' ? { urgency } : {}),
+        ...(bloodGroup && bloodGroup !== "all"
+          ? { bloodGroup: bloodGroup.toUpperCase() }
+          : {}),
+        ...(district && district !== "all"
+          ? { recipientDistrict: district }
+          : {}),
+        ...(upazila && upazila !== "all" ? { recipientUpazila: upazila } : {}),
+        ...(urgency && urgency !== "all" ? { urgency } : {}),
       },
     },
     {
@@ -413,10 +445,12 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
         bloodGroupStats: [
           {
             $group: {
-              _id: '$bloodGroup',
+              _id: "$bloodGroup",
               count: { $sum: 1 },
               urgent: {
-                $sum: { $cond: [{ $in: ['$urgency', ['high', 'critical']] }, 1, 0] },
+                $sum: {
+                  $cond: [{ $in: ["$urgency", ["high", "critical"]] }, 1, 0],
+                },
               },
             },
           },
@@ -426,8 +460,8 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
           {
             $group: {
               _id: {
-                district: '$recipientDistrict',
-                upazila: '$recipientUpazila',
+                district: "$recipientDistrict",
+                upazila: "$recipientUpazila",
               },
               count: { $sum: 1 },
             },
@@ -438,7 +472,7 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
         urgencyStats: [
           {
             $group: {
-              _id: '$urgency',
+              _id: "$urgency",
               count: { $sum: 1 },
             },
           },
@@ -447,7 +481,7 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
         hospitalStats: [
           {
             $group: {
-              _id: '$hospitalName',
+              _id: "$hospitalName",
               count: { $sum: 1 },
             },
           },
@@ -465,14 +499,16 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
       userName: req.user.name,
       userEmail: req.user.email,
       userRole: req.user.role,
-      action: 'Searched for Donation Requests',
-      actionType: 'search',
-      category: 'search',
-      description: `Searched donation requests with filters: Blood Group=${bloodGroup || 'Any'}, District=${district || 'Any'}, Status=${status}`,
+      action: "Searched for Donation Requests",
+      actionType: "search",
+      category: "search",
+      description: `Searched donation requests with filters: Blood Group=${
+        bloodGroup || "Any"
+      }, District=${district || "Any"}, Status=${status}`,
       details: `Found ${total} requests matching criteria`,
-      status: 'success',
+      status: "success",
       userIp: req.ip,
-      userAgent: req.headers['user-agent'],
+      userAgent: req.headers["user-agent"],
     });
   }
 
@@ -481,11 +517,11 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
     count: requests.length,
     total,
     filters: {
-      bloodGroup: bloodGroup || 'Any',
-      district: district || 'Any',
-      upazila: upazila || 'Any',
+      bloodGroup: bloodGroup || "Any",
+      district: district || "Any",
+      upazila: upazila || "Any",
       status,
-      urgency: urgency || 'Any',
+      urgency: urgency || "Any",
       sortBy,
       sortOrder,
     },
@@ -504,32 +540,33 @@ exports.searchDonationRequests = asyncHandler(async (req, res, next) => {
 // @desc    Global search across all entities
 // @route   GET /api/search/global
 // @access  Private (based on role)
-exports.globalSearch = asyncHandler(async (req, res, next) => {
-  const { q: searchQuery, type = 'all', page = 1, limit = 10 } = req.query;
+export const globalSearch = asyncHandler(async (req, res, next) => {
+  const { q: searchQuery, type = "all", page = 1, limit = 10 } = req.query;
 
   if (!searchQuery || searchQuery.trim().length < 2) {
-    return next(new ErrorResponse('Search query must be at least 2 characters', 400));
+    return next(
+      new ErrorResponse("Search query must be at least 2 characters", 400)
+    );
   }
 
   const pageNum = parseInt(page, 10);
   const limitNum = parseInt(limit, 10);
   const skip = (pageNum - 1) * limitNum;
 
-  const searchRegex = { $regex: searchQuery, $options: 'i' };
+  const searchRegex = { $regex: searchQuery, $options: "i" };
   let results = [];
   let total = 0;
   let entityCounts = {};
 
   // Determine what to search based on user role and type parameter
-  const searchTypes = type === 'all' 
-    ? ['users', 'donations', 'fundings', 'contacts']
-    : [type];
+  const searchTypes =
+    type === "all" ? ["users", "donations", "fundings", "contacts"] : [type];
 
   // Build search promises based on types
   const searchPromises = [];
 
   // Search Users
-  if (searchTypes.includes('users')) {
+  if (searchTypes.includes("users")) {
     const userFilter = {
       $or: [
         { name: searchRegex },
@@ -539,25 +576,31 @@ exports.globalSearch = asyncHandler(async (req, res, next) => {
     };
 
     // Role-based filtering for users
-    if (req.user.role === 'donor') {
-      userFilter.role = 'donor'; // Donors can only see other donors
+    if (req.user.role === "donor") {
+      userFilter.role = "donor"; // Donors can only see other donors
     }
 
     searchPromises.push(
       User.find(userFilter)
-        .select('-password')
-        .skip(type === 'users' ? skip : 0)
-        .limit(type === 'users' ? limitNum : 5)
-        .then(users => ({
-          type: 'users',
+        .select("-password")
+        .skip(type === "users" ? skip : 0)
+        .limit(type === "users" ? limitNum : 5)
+        .then((users) => ({
+          type: "users",
           data: users,
-          total: type === 'users' ? User.countDocuments(userFilter) : Promise.resolve(users.length),
+          total:
+            type === "users"
+              ? User.countDocuments(userFilter)
+              : Promise.resolve(users.length),
         }))
     );
   }
 
   // Search Donation Requests
-  if (searchTypes.includes('donations') && (req.user.role === 'admin' || req.user.role === 'volunteer')) {
+  if (
+    searchTypes.includes("donations") &&
+    (req.user.role === "admin" || req.user.role === "volunteer")
+  ) {
     const donationFilter = {
       isActive: true,
       $or: [
@@ -571,20 +614,26 @@ exports.globalSearch = asyncHandler(async (req, res, next) => {
 
     searchPromises.push(
       DonationRequest.find(donationFilter)
-        .populate('requester', 'name email')
-        .populate('donor', 'name email')
-        .skip(type === 'donations' ? skip : 0)
-        .limit(type === 'donations' ? limitNum : 5)
-        .then(donations => ({
-          type: 'donations',
+        .populate("requester", "name email")
+        .populate("donor", "name email")
+        .skip(type === "donations" ? skip : 0)
+        .limit(type === "donations" ? limitNum : 5)
+        .then((donations) => ({
+          type: "donations",
           data: donations,
-          total: type === 'donations' ? DonationRequest.countDocuments(donationFilter) : Promise.resolve(donations.length),
+          total:
+            type === "donations"
+              ? DonationRequest.countDocuments(donationFilter)
+              : Promise.resolve(donations.length),
         }))
     );
   }
 
   // Search Fundings (admin/volunteer only)
-  if (searchTypes.includes('fundings') && (req.user.role === 'admin' || req.user.role === 'volunteer')) {
+  if (
+    searchTypes.includes("fundings") &&
+    (req.user.role === "admin" || req.user.role === "volunteer")
+  ) {
     const fundingFilter = {
       $or: [
         { donorName: searchRegex },
@@ -596,19 +645,25 @@ exports.globalSearch = asyncHandler(async (req, res, next) => {
 
     searchPromises.push(
       Funding.find(fundingFilter)
-        .populate('donor', 'name email')
-        .skip(type === 'fundings' ? skip : 0)
-        .limit(type === 'fundings' ? limitNum : 5)
-        .then(fundings => ({
-          type: 'fundings',
+        .populate("donor", "name email")
+        .skip(type === "fundings" ? skip : 0)
+        .limit(type === "fundings" ? limitNum : 5)
+        .then((fundings) => ({
+          type: "fundings",
           data: fundings,
-          total: type === 'fundings' ? Funding.countDocuments(fundingFilter) : Promise.resolve(fundings.length),
+          total:
+            type === "fundings"
+              ? Funding.countDocuments(fundingFilter)
+              : Promise.resolve(fundings.length),
         }))
     );
   }
 
   // Search Contacts (admin/volunteer only)
-  if (searchTypes.includes('contacts') && (req.user.role === 'admin' || req.user.role === 'volunteer')) {
+  if (
+    searchTypes.includes("contacts") &&
+    (req.user.role === "admin" || req.user.role === "volunteer")
+  ) {
     const contactFilter = {
       $or: [
         { name: searchRegex },
@@ -620,28 +675,35 @@ exports.globalSearch = asyncHandler(async (req, res, next) => {
 
     searchPromises.push(
       Contact.find(contactFilter)
-        .populate('user', 'name email')
-        .populate('assignedTo', 'name email')
-        .skip(type === 'contacts' ? skip : 0)
-        .limit(type === 'contacts' ? limitNum : 5)
-        .then(contacts => ({
-          type: 'contacts',
+        .populate("user", "name email")
+        .populate("assignedTo", "name email")
+        .skip(type === "contacts" ? skip : 0)
+        .limit(type === "contacts" ? limitNum : 5)
+        .then((contacts) => ({
+          type: "contacts",
           data: contacts,
-          total: type === 'contacts' ? Contact.countDocuments(contactFilter) : Promise.resolve(contacts.length),
+          total:
+            type === "contacts"
+              ? Contact.countDocuments(contactFilter)
+              : Promise.resolve(contacts.length),
         }))
     );
   }
 
   // Execute all search promises
-  const searchResults = await Promise.all(searchPromises.map(p => p.then(async result => ({
-    type: result.type,
-    data: result.data,
-    total: await result.total,
-  }))));
+  const searchResults = await Promise.all(
+    searchPromises.map((p) =>
+      p.then(async (result) => ({
+        type: result.type,
+        data: result.data,
+        total: await result.total,
+      }))
+    )
+  );
 
   // Combine results
-  results = searchResults.flatMap(result => 
-    result.data.map(item => ({
+  results = searchResults.flatMap((result) =>
+    result.data.map((item) => ({
       ...item.toObject(),
       _type: result.type.slice(0, -1), // Remove 's' for singular
       type: result.type,
@@ -659,22 +721,22 @@ exports.globalSearch = asyncHandler(async (req, res, next) => {
   // Sort results by relevance
   results.sort((a, b) => {
     // Priority: exact matches in name/title first
-    const aName = a.name || a.recipientName || a.donorName || '';
-    const bName = b.name || b.recipientName || b.donorName || '';
-    
+    const aName = a.name || a.recipientName || a.donorName || "";
+    const bName = b.name || b.recipientName || b.donorName || "";
+
     const aExactMatch = aName.toLowerCase() === searchQuery.toLowerCase();
     const bExactMatch = bName.toLowerCase() === searchQuery.toLowerCase();
-    
+
     if (aExactMatch && !bExactMatch) return -1;
     if (!aExactMatch && bExactMatch) return 1;
-    
+
     // Then by type priority
     const typePriority = { users: 1, donations: 2, fundings: 3, contacts: 4 };
     return typePriority[a.type] - typePriority[b.type];
   });
 
   // Apply pagination if searching all types
-  if (type === 'all') {
+  if (type === "all") {
     results = results.slice(skip, skip + limitNum);
   }
 
@@ -684,14 +746,14 @@ exports.globalSearch = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Performed Global Search',
-    actionType: 'search',
-    category: 'search',
+    action: "Performed Global Search",
+    actionType: "search",
+    category: "search",
     description: `Global search for "${searchQuery}"`,
-    details: `Types: ${searchTypes.join(', ')}, Found ${total} results`,
-    status: 'success',
+    details: `Types: ${searchTypes.join(", ")}, Found ${total} results`,
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
     request: {
       method: req.method,
       url: req.originalUrl,
@@ -710,13 +772,16 @@ exports.globalSearch = asyncHandler(async (req, res, next) => {
       page: pageNum,
       limit: limitNum,
     },
-    pagination: type === 'all' ? {
-      page: pageNum,
-      limit: limitNum,
-      totalPages: Math.ceil(total / limitNum),
-      hasNextPage: pageNum * limitNum < total,
-      hasPrevPage: pageNum > 1,
-    } : null,
+    pagination:
+      type === "all"
+        ? {
+            page: pageNum,
+            limit: limitNum,
+            totalPages: Math.ceil(total / limitNum),
+            hasNextPage: pageNum * limitNum < total,
+            hasPrevPage: pageNum > 1,
+          }
+        : null,
     data: results,
   });
 });
@@ -724,35 +789,35 @@ exports.globalSearch = asyncHandler(async (req, res, next) => {
 // @desc    Get available filters for search
 // @route   GET /api/search/filters
 // @access  Public
-exports.getSearchFilters = asyncHandler(async (req, res, next) => {
-  const { type = 'donors' } = req.query;
+export const getSearchFilters = asyncHandler(async (req, res, next) => {
+  const { type = "donors" } = req.query;
 
   let filters = {};
 
   switch (type) {
-    case 'donors':
+    case "donors":
       // Get unique blood groups from donors
-      const bloodGroups = await User.distinct('bloodGroup', {
-        role: 'donor',
-        status: 'active',
+      const bloodGroups = await User.distinct("bloodGroup", {
+        role: "donor",
+        status: "active",
         bloodGroup: { $ne: null },
       });
 
       // Get unique districts from donors
-      const donorDistricts = await User.distinct('district', {
-        role: 'donor',
-        status: 'active',
-        district: { $ne: null, $ne: '' },
+      const donorDistricts = await User.distinct("district", {
+        role: "donor",
+        status: "active",
+        district: { $ne: null, $ne: "" },
       });
 
       // Get upazilas based on selected district
       let donorUpazilas = [];
       if (req.query.district) {
-        donorUpazilas = await User.distinct('upazila', {
-          role: 'donor',
-          status: 'active',
+        donorUpazilas = await User.distinct("upazila", {
+          role: "donor",
+          status: "active",
           district: req.query.district,
-          upazila: { $ne: null, $ne: '' },
+          upazila: { $ne: null, $ne: "" },
         });
       }
 
@@ -761,42 +826,45 @@ exports.getSearchFilters = asyncHandler(async (req, res, next) => {
         districts: donorDistricts.sort(),
         upazilas: donorUpazilas.sort(),
         sortOptions: [
-          { value: 'relevance', label: 'Most Relevant' },
-          { value: 'recent', label: 'Most Recent' },
-          { value: 'donations', label: 'Most Donations' },
-          { value: 'name', label: 'Name (A-Z)' },
-          { value: 'location', label: 'Location' },
+          { value: "relevance", label: "Most Relevant" },
+          { value: "recent", label: "Most Recent" },
+          { value: "donations", label: "Most Donations" },
+          { value: "name", label: "Name (A-Z)" },
+          { value: "location", label: "Location" },
         ],
         availabilityOptions: [
-          { value: 'true', label: 'Available Only' },
-          { value: 'false', label: 'All Donors' },
+          { value: "true", label: "Available Only" },
+          { value: "false", label: "All Donors" },
         ],
       };
       break;
 
-    case 'donation-requests':
+    case "donation-requests":
       // Get unique blood groups from donation requests
-      const requestBloodGroups = await DonationRequest.distinct('bloodGroup', {
+      const requestBloodGroups = await DonationRequest.distinct("bloodGroup", {
         isActive: true,
-        status: 'pending',
+        status: "pending",
         bloodGroup: { $ne: null },
       });
 
       // Get unique districts from donation requests
-      const requestDistricts = await DonationRequest.distinct('recipientDistrict', {
-        isActive: true,
-        status: 'pending',
-        recipientDistrict: { $ne: null, $ne: '' },
-      });
+      const requestDistricts = await DonationRequest.distinct(
+        "recipientDistrict",
+        {
+          isActive: true,
+          status: "pending",
+          recipientDistrict: { $ne: null, $ne: "" },
+        }
+      );
 
       // Get upazilas based on selected district
       let requestUpazilas = [];
       if (req.query.district) {
-        requestUpazilas = await DonationRequest.distinct('recipientUpazila', {
+        requestUpazilas = await DonationRequest.distinct("recipientUpazila", {
           isActive: true,
-          status: 'pending',
+          status: "pending",
           recipientDistrict: req.query.district,
-          recipientUpazila: { $ne: null, $ne: '' },
+          recipientUpazila: { $ne: null, $ne: "" },
         });
       }
 
@@ -805,46 +873,46 @@ exports.getSearchFilters = asyncHandler(async (req, res, next) => {
         districts: requestDistricts.sort(),
         upazilas: requestUpazilas.sort(),
         statusOptions: [
-          { value: 'pending', label: 'Pending' },
-          { value: 'inprogress', label: 'In Progress' },
-          { value: 'all', label: 'All Statuses' },
+          { value: "pending", label: "Pending" },
+          { value: "inprogress", label: "In Progress" },
+          { value: "all", label: "All Statuses" },
         ],
         urgencyOptions: [
-          { value: 'all', label: 'All Urgency Levels' },
-          { value: 'critical', label: 'Critical' },
-          { value: 'high', label: 'High' },
-          { value: 'medium', label: 'Medium' },
-          { value: 'low', label: 'Low' },
+          { value: "all", label: "All Urgency Levels" },
+          { value: "critical", label: "Critical" },
+          { value: "high", label: "High" },
+          { value: "medium", label: "Medium" },
+          { value: "low", label: "Low" },
         ],
         sortOptions: [
-          { value: 'urgency', label: 'Urgency (High to Low)' },
-          { value: 'date', label: 'Date (Soonest First)' },
-          { value: 'created', label: 'Recently Created' },
-          { value: 'hospital', label: 'Hospital Name' },
+          { value: "urgency", label: "Urgency (High to Low)" },
+          { value: "date", label: "Date (Soonest First)" },
+          { value: "created", label: "Recently Created" },
+          { value: "hospital", label: "Hospital Name" },
         ],
       };
       break;
 
-    case 'global':
+    case "global":
       filters = {
         searchTypes: [
-          { value: 'all', label: 'All' },
-          { value: 'users', label: 'Users' },
-          { value: 'donations', label: 'Donation Requests' },
-          { value: 'fundings', label: 'Donations' },
-          { value: 'contacts', label: 'Contacts' },
+          { value: "all", label: "All" },
+          { value: "users", label: "Users" },
+          { value: "donations", label: "Donation Requests" },
+          { value: "fundings", label: "Donations" },
+          { value: "contacts", label: "Contacts" },
         ],
         minQueryLength: 2,
       };
-      
+
       // Add role-specific filters
       if (req.user) {
-        if (req.user.role === 'admin' || req.user.role === 'volunteer') {
+        if (req.user.role === "admin" || req.user.role === "volunteer") {
           filters.adminFilters = {
             quickFilters: [
-              { value: 'urgent', label: 'Urgent Requests' },
-              { value: 'unverified', label: 'Unverified Donations' },
-              { value: 'blocked', label: 'Blocked Users' },
+              { value: "urgent", label: "Urgent Requests" },
+              { value: "unverified", label: "Unverified Donations" },
+              { value: "blocked", label: "Blocked Users" },
             ],
           };
         }
@@ -852,13 +920,17 @@ exports.getSearchFilters = asyncHandler(async (req, res, next) => {
       break;
 
     default:
-      return next(new ErrorResponse('Invalid search type', 400));
+      return next(new ErrorResponse("Invalid search type", 400));
   }
 
   // Add common location data
   filters.locationData = {
-    allDistricts: districts.map(d => ({ value: d.name, label: d.name })),
-    allUpazilas: upazilas.map(u => ({ value: u.name, label: u.name, district: u.district })),
+    allDistricts: districts.map((d) => ({ value: d.name, label: d.name })),
+    allUpazilas: upazilas.map((u) => ({
+      value: u.name,
+      label: u.name,
+      district: u.district,
+    })),
   };
 
   res.status(200).json({
@@ -871,36 +943,38 @@ exports.getSearchFilters = asyncHandler(async (req, res, next) => {
 // @desc    Quick search for urgent needs
 // @route   GET /api/search/urgent
 // @access  Public
-exports.getUrgentNeeds = asyncHandler(async (req, res, next) => {
+export const getUrgentNeeds = asyncHandler(async (req, res, next) => {
   const limit = parseInt(req.query.limit, 10) || 5;
 
   // Get urgent donation requests
   const urgentRequests = await DonationRequest.find({
     isActive: true,
-    status: 'pending',
-    urgency: { $in: ['high', 'critical'] },
+    status: "pending",
+    urgency: { $in: ["high", "critical"] },
     donationDate: { $gte: new Date() }, // Not expired
   })
-    .populate('requester', 'name email avatar')
-    .select('recipientName bloodGroup hospitalName donationDate urgency recipientDistrict recipientUpazila')
+    .populate("requester", "name email avatar")
+    .select(
+      "recipientName bloodGroup hospitalName donationDate urgency recipientDistrict recipientUpazila"
+    )
     .sort({ urgency: -1, donationDate: 1 })
     .limit(limit);
 
   // Get rare blood group needs
-  const rareBloodGroups = ['AB-', 'B-', 'A-', 'O-'];
+  const rareBloodGroups = ["AB-", "B-", "A-", "O-"];
   const rareBloodNeeds = await DonationRequest.find({
     isActive: true,
-    status: 'pending',
+    status: "pending",
     bloodGroup: { $in: rareBloodGroups },
     donationDate: { $gte: new Date() },
   })
-    .select('bloodGroup recipientDistrict recipientUpazila donationDate')
+    .select("bloodGroup recipientDistrict recipientUpazila donationDate")
     .sort({ donationDate: 1 })
     .limit(limit);
 
   // Group rare blood needs by type
   const rareBloodStats = {};
-  rareBloodNeeds.forEach(need => {
+  rareBloodNeeds.forEach((need) => {
     if (!rareBloodStats[need.bloodGroup]) {
       rareBloodStats[need.bloodGroup] = [];
     }
@@ -915,19 +989,19 @@ exports.getUrgentNeeds = asyncHandler(async (req, res, next) => {
     {
       $match: {
         isActive: true,
-        status: 'pending',
-        urgency: { $in: ['high', 'critical'] },
+        status: "pending",
+        urgency: { $in: ["high", "critical"] },
         donationDate: { $gte: new Date() },
       },
     },
     {
       $group: {
         _id: {
-          district: '$recipientDistrict',
-          upazila: '$recipientUpazila',
+          district: "$recipientDistrict",
+          upazila: "$recipientUpazila",
         },
         count: { $sum: 1 },
-        bloodGroups: { $addToSet: '$bloodGroup' },
+        bloodGroups: { $addToSet: "$bloodGroup" },
       },
     },
     { $sort: { count: -1 } },
@@ -952,11 +1026,11 @@ exports.getUrgentNeeds = asyncHandler(async (req, res, next) => {
 // @desc    Export search results as PDF
 // @route   POST /api/search/export
 // @access  Private
-exports.exportSearchResults = asyncHandler(async (req, res, next) => {
-  const { type, filters, format = 'pdf' } = req.body;
+export const exportSearchResults = asyncHandler(async (req, res, next) => {
+  const { type, filters, format = "pdf" } = req.body;
 
   if (!type || !filters) {
-    return next(new ErrorResponse('Type and filters are required', 400));
+    return next(new ErrorResponse("Type and filters are required", 400));
   }
 
   let data;
@@ -964,20 +1038,20 @@ exports.exportSearchResults = asyncHandler(async (req, res, next) => {
   let exportData;
 
   switch (type) {
-    case 'donors':
+    case "donors":
       // Apply filters and get donors
       const donorFilter = {
-        role: 'donor',
-        status: 'active',
+        role: "donor",
+        status: "active",
       };
 
-      if (filters.bloodGroup && filters.bloodGroup !== 'all') {
+      if (filters.bloodGroup && filters.bloodGroup !== "all") {
         donorFilter.bloodGroup = filters.bloodGroup.toUpperCase();
       }
-      if (filters.district && filters.district !== 'all') {
+      if (filters.district && filters.district !== "all") {
         donorFilter.district = filters.district;
       }
-      if (filters.upazila && filters.upazila !== 'all') {
+      if (filters.upazila && filters.upazila !== "all") {
         donorFilter.upazila = filters.upazila;
       }
       if (filters.availableOnly) {
@@ -985,68 +1059,76 @@ exports.exportSearchResults = asyncHandler(async (req, res, next) => {
       }
 
       data = await User.find(donorFilter)
-        .select('name email bloodGroup district upazila phone totalDonations lastDonationDate isAvailable')
+        .select(
+          "name email bloodGroup district upazila phone totalDonations lastDonationDate isAvailable"
+        )
         .sort({ name: 1 });
 
-      exportData = data.map(donor => ({
+      exportData = data.map((donor) => ({
         Name: donor.name,
         Email: donor.email,
-        'Blood Group': donor.bloodGroup,
+        "Blood Group": donor.bloodGroup,
         District: donor.district,
         Upazila: donor.upazila,
-        Phone: donor.phone || 'N/A',
-        'Total Donations': donor.totalDonations || 0,
-        'Last Donation': donor.lastDonationDate 
-          ? new Date(donor.lastDonationDate).toLocaleDateString() 
-          : 'Never',
-        'Currently Available': donor.isAvailable ? 'Yes' : 'No',
+        Phone: donor.phone || "N/A",
+        "Total Donations": donor.totalDonations || 0,
+        "Last Donation": donor.lastDonationDate
+          ? new Date(donor.lastDonationDate).toLocaleDateString()
+          : "Never",
+        "Currently Available": donor.isAvailable ? "Yes" : "No",
       }));
 
-      fileName = `donors_${filters.bloodGroup || 'all'}_${filters.district || 'all'}_${new Date().toISOString().split('T')[0]}`;
+      fileName = `donors_${filters.bloodGroup || "all"}_${
+        filters.district || "all"
+      }_${new Date().toISOString().split("T")[0]}`;
       break;
 
-    case 'donation-requests':
+    case "donation-requests":
       const requestFilter = {
         isActive: true,
-        status: filters.status || 'pending',
+        status: filters.status || "pending",
       };
 
-      if (filters.bloodGroup && filters.bloodGroup !== 'all') {
+      if (filters.bloodGroup && filters.bloodGroup !== "all") {
         requestFilter.bloodGroup = filters.bloodGroup.toUpperCase();
       }
-      if (filters.district && filters.district !== 'all') {
+      if (filters.district && filters.district !== "all") {
         requestFilter.recipientDistrict = filters.district;
       }
-      if (filters.upazila && filters.upazila !== 'all') {
+      if (filters.upazila && filters.upazila !== "all") {
         requestFilter.recipientUpazila = filters.upazila;
       }
-      if (filters.urgency && filters.urgency !== 'all') {
+      if (filters.urgency && filters.urgency !== "all") {
         requestFilter.urgency = filters.urgency;
       }
 
       data = await DonationRequest.find(requestFilter)
-        .populate('requester', 'name email phone')
-        .select('recipientName bloodGroup hospitalName hospitalAddress donationDate donationTime urgency status')
+        .populate("requester", "name email phone")
+        .select(
+          "recipientName bloodGroup hospitalName hospitalAddress donationDate donationTime urgency status"
+        )
         .sort({ donationDate: 1 });
 
-      exportData = data.map(request => ({
-        'Patient Name': request.recipientName,
-        'Blood Group': request.bloodGroup,
+      exportData = data.map((request) => ({
+        "Patient Name": request.recipientName,
+        "Blood Group": request.bloodGroup,
         Hospital: request.hospitalName,
         Address: request.hospitalAddress,
-        'Donation Date': new Date(request.donationDate).toLocaleDateString(),
-        'Donation Time': request.donationTime,
+        "Donation Date": new Date(request.donationDate).toLocaleDateString(),
+        "Donation Time": request.donationTime,
         Urgency: request.urgency,
         Status: request.status,
-        'Requester Name': request.requester?.name || 'N/A',
-        'Requester Contact': request.requester?.phone || 'N/A',
+        "Requester Name": request.requester?.name || "N/A",
+        "Requester Contact": request.requester?.phone || "N/A",
       }));
 
-      fileName = `donation_requests_${filters.status || 'pending'}_${new Date().toISOString().split('T')[0]}`;
+      fileName = `donation_requests_${filters.status || "pending"}_${
+        new Date().toISOString().split("T")[0]
+      }`;
       break;
 
     default:
-      return next(new ErrorResponse('Invalid export type', 400));
+      return next(new ErrorResponse("Invalid export type", 400));
   }
 
   // Log export activity
@@ -1055,23 +1137,25 @@ exports.exportSearchResults = asyncHandler(async (req, res, next) => {
     userName: req.user.name,
     userEmail: req.user.email,
     userRole: req.user.role,
-    action: 'Exported Search Results',
-    actionType: 'read',
-    category: 'search',
+    action: "Exported Search Results",
+    actionType: "read",
+    category: "search",
     description: `Exported ${type} search results`,
-    details: `Filters: ${JSON.stringify(filters)}, Format: ${format}, Records: ${exportData.length}`,
-    status: 'success',
+    details: `Filters: ${JSON.stringify(
+      filters
+    )}, Format: ${format}, Records: ${exportData.length}`,
+    status: "success",
     userIp: req.ip,
-    userAgent: req.headers['user-agent'],
+    userAgent: req.headers["user-agent"],
   });
 
-  if (format === 'csv') {
+  if (format === "csv") {
     // Convert to CSV
-    const { Parser } = require('json2csv');
+    const { Parser } = require("json2csv");
     const json2csvParser = new Parser();
     const csv = json2csvParser.parse(exportData);
 
-    res.header('Content-Type', 'text/csv');
+    res.header("Content-Type", "text/csv");
     res.attachment(`${fileName}.csv`);
     return res.send(csv);
   }
