@@ -1,9 +1,9 @@
 const errorHandler = (err, req, res, next) => {
-  console.error('Error:', {
-      message: err.message,
-      stack: process.env.NODE_ENV === 'development' ? err.stack : undefined,
-      path: req.path,
-      method: req.method
+  console.error("Error:", {
+    message: err.message,
+    stack: process.env.NODE_ENV === "development" ? err.stack : undefined,
+    path: req.path,
+    method: req.method,
   });
 
   let statusCode = res.statusCode === 200 ? 500 : res.statusCode;
@@ -11,65 +11,67 @@ const errorHandler = (err, req, res, next) => {
   let errors = null;
 
   // Mongoose validation error
-  if (err.name === 'ValidationError') {
-      statusCode = 400;
-      message = 'Validation failed';
-      errors = Object.values(err.errors).map(error => ({
-          field: error.path,
-          message: error.message
-      }));
+  if (err.name === "ValidationError") {
+    statusCode = 400;
+    message = "Validation failed";
+    errors = Object.values(err.errors).map((error) => ({
+      field: error.path,
+      message: error.message,
+    }));
   }
 
   // Mongoose duplicate key error
   if (err.code === 11000) {
-      statusCode = 400;
-      message = 'Duplicate field value entered';
-      const field = Object.keys(err.keyPattern)[0];
-      errors = [{
-          field,
-          message: `${field} already exists`
-      }];
+    statusCode = 400;
+    message = "Duplicate field value entered";
+    const field = Object.keys(err.keyPattern)[0];
+    errors = [
+      {
+        field,
+        message: `${field} already exists`,
+      },
+    ];
   }
 
   // Mongoose CastError (invalid ObjectId)
-  if (err.name === 'CastError') {
-      statusCode = 400;
-      message = 'Invalid resource ID';
+  if (err.name === "CastError") {
+    statusCode = 400;
+    message = "Invalid resource ID";
   }
 
   // JWT errors
-  if (err.name === 'JsonWebTokenError') {
-      statusCode = 401;
-      message = 'Invalid token';
+  if (err.name === "JsonWebTokenError") {
+    statusCode = 401;
+    message = "Invalid token";
   }
 
-  if (err.name === 'TokenExpiredError') {
-      statusCode = 401;
-      message = 'Token expired';
+  if (err.name === "TokenExpiredError") {
+    statusCode = 401;
+    message = "Token expired";
   }
 
   // Multer errors
-  if (err.name === 'MulterError') {
-      statusCode = 400;
-      if (err.code === 'LIMIT_FILE_SIZE') {
-          message = 'File too large';
-      } else if (err.code === 'LIMIT_FILE_COUNT') {
-          message = 'Too many files';
-      } else if (err.code === 'LIMIT_UNEXPECTED_FILE') {
-          message = 'Invalid file field';
-      }
+  if (err.name === "MulterError") {
+    statusCode = 400;
+    if (err.code === "LIMIT_FILE_SIZE") {
+      message = "File too large";
+    } else if (err.code === "LIMIT_FILE_COUNT") {
+      message = "Too many files";
+    } else if (err.code === "LIMIT_UNEXPECTED_FILE") {
+      message = "Invalid file field";
+    }
   }
 
   // Custom error with status code
   if (err.statusCode) {
-      statusCode = err.statusCode;
+    statusCode = err.statusCode;
   }
 
   res.status(statusCode).json({
-      success: false,
-      message,
-      errors,
-      stack: process.env.NODE_ENV === 'development' ? err.stack : undefined
+    success: false,
+    message,
+    errors,
+    stack: process.env.NODE_ENV === "development" ? err.stack : undefined,
   });
 };
 
@@ -83,8 +85,8 @@ const asyncHandler = (fn) => (req, res, next) => {
   Promise.resolve(fn(req, res, next)).catch(next);
 };
 
-module.exports = {
+export default {
   errorHandler,
   notFound,
-  asyncHandler
+  asyncHandler,
 };
