@@ -1,144 +1,148 @@
-const multer = require('multer');
-const path = require('path');
-const fs = require('fs');
+import multer from "multer";
+import path from "path";
+import fs from "fs";
 
 // Ensure uploads directory exists
-const uploadDir = 'uploads/';
+const uploadDir = "uploads/";
 if (!fs.existsSync(uploadDir)) {
-    fs.mkdirSync(uploadDir, { recursive: true });
+  fs.mkdirSync(uploadDir, { recursive: true });
 }
 
 // Configure storage
 const storage = multer.diskStorage({
-    destination: (req, file, cb) => {
-        cb(null, uploadDir);
-    },
-    filename: (req, file, cb) => {
-        const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
-        const ext = path.extname(file.originalname).toLowerCase();
-        cb(null, file.fieldname + '-' + uniqueSuffix + ext);
-    }
+  destination: (req, file, cb) => {
+    cb(null, uploadDir);
+  },
+  filename: (req, file, cb) => {
+    const uniqueSuffix = Date.now() + "-" + Math.round(Math.random() * 1e9);
+    const ext = path.extname(file.originalname).toLowerCase();
+    cb(null, file.fieldname + "-" + uniqueSuffix + ext);
+  },
 });
 
 // File filter
 const fileFilter = (req, file, cb) => {
-    const allowedTypes = /jpeg|jpg|png|gif|pdf/;
-    const extname = allowedTypes.test(path.extname(file.originalname).toLowerCase());
-    const mimetype = allowedTypes.test(file.mimetype);
+  const allowedTypes = /jpeg|jpg|png|gif|pdf/;
+  const extname = allowedTypes.test(
+    path.extname(file.originalname).toLowerCase()
+  );
+  const mimetype = allowedTypes.test(file.mimetype);
 
-    if (mimetype && extname) {
-        return cb(null, true);
-    } else {
-        cb(new Error('Only image files (JPEG, JPG, PNG, GIF) and PDF are allowed'));
-    }
+  if (mimetype && extname) {
+    return cb(null, true);
+  } else {
+    cb(new Error("Only image files (JPEG, JPG, PNG, GIF) and PDF are allowed"));
+  }
 };
 
 // Initialize upload
 const upload = multer({
-    storage: storage,
-    limits: {
-        fileSize: 5 * 1024 * 1024, // 5MB
-        files: 1
-    },
-    fileFilter: fileFilter
+  storage: storage,
+  limits: {
+    fileSize: 5 * 1024 * 1024, // 5MB
+    files: 1,
+  },
+  fileFilter: fileFilter,
 });
 
 // Single file upload middleware
 const uploadSingle = (fieldName) => {
-    return (req, res, next) => {
-        upload.single(fieldName)(req, res, (err) => {
-            if (err instanceof multer.MulterError) {
-                return res.status(400).json({
-                    success: false,
-                    message: err.message
-                });
-            } else if (err) {
-                return res.status(400).json({
-                    success: false,
-                    message: err.message
-                });
-            }
-            next();
+  return (req, res, next) => {
+    upload.single(fieldName)(req, res, (err) => {
+      if (err instanceof multer.MulterError) {
+        return res.status(400).json({
+          success: false,
+          message: err.message,
         });
-    };
+      } else if (err) {
+        return res.status(400).json({
+          success: false,
+          message: err.message,
+        });
+      }
+      next();
+    });
+  };
 };
 
 // Multiple files upload middleware
 const uploadMultiple = (fieldName, maxCount = 5) => {
-    return (req, res, next) => {
-        upload.array(fieldName, maxCount)(req, res, (err) => {
-            if (err instanceof multer.MulterError) {
-                return res.status(400).json({
-                    success: false,
-                    message: err.message
-                });
-            } else if (err) {
-                return res.status(400).json({
-                    success: false,
-                    message: err.message
-                });
-            }
-            next();
+  return (req, res, next) => {
+    upload.array(fieldName, maxCount)(req, res, (err) => {
+      if (err instanceof multer.MulterError) {
+        return res.status(400).json({
+          success: false,
+          message: err.message,
         });
-    };
+      } else if (err) {
+        return res.status(400).json({
+          success: false,
+          message: err.message,
+        });
+      }
+      next();
+    });
+  };
 };
 
 // Specific fields upload middleware
 const uploadFields = (fields) => {
-    return (req, res, next) => {
-        upload.fields(fields)(req, res, (err) => {
-            if (err instanceof multer.MulterError) {
-                return res.status(400).json({
-                    success: false,
-                    message: err.message
-                });
-            } else if (err) {
-                return res.status(400).json({
-                    success: false,
-                    message: err.message
-                });
-            }
-            next();
+  return (req, res, next) => {
+    upload.fields(fields)(req, res, (err) => {
+      if (err instanceof multer.MulterError) {
+        return res.status(400).json({
+          success: false,
+          message: err.message,
         });
-    };
+      } else if (err) {
+        return res.status(400).json({
+          success: false,
+          message: err.message,
+        });
+      }
+      next();
+    });
+  };
 };
 
 // Clean up uploaded files on error
 const cleanupUploads = (req, res, next) => {
-    const originalSend = res.send;
-    
-    res.send = function(data) {
-        if (res.statusCode >= 400 && req.file) {
-            const filePath = req.file.path;
-            if (fs.existsSync(filePath)) {
-                fs.unlink(filePath, (err) => {
-                    if (err) console.error('Error deleting file:', err);
-                });
-            }
+  const originalSend = res.send;
+
+  res.send = function (data) {
+    if (res.statusCode >= 400 && req.file) {
+      const filePath = req.file.path;
+      if (fs.existsSync(filePath)) {
+        fs.unlink(filePath, (err) => {
+          if (err) console.error("Error deleting file:", err);
+        });
+      }
+    }
+
+    if (res.statusCode >= 400 && req.files) {
+      const files = Array.isArray(req.files)
+        ? req.files
+        : Object.values(req.files).flat();
+      files.forEach((file) => {
+        if (fs.existsSync(file.path)) {
+          fs.unlink(file.path, (err) => {
+            if (err) console.error("Error deleting file:", err);
+          });
         }
-        
-        if (res.statusCode >= 400 && req.files) {
-            const files = Array.isArray(req.files) ? req.files : Object.values(req.files).flat();
-            files.forEach(file => {
-                if (fs.existsSync(file.path)) {
-                    fs.unlink(file.path, (err) => {
-                        if (err) console.error('Error deleting file:', err);
-                    });
-                }
-            });
-        }
-        
-        return originalSend.call(this, data);
-    };
-    
-    next();
+      });
+    }
+
+    return originalSend.call(this, data);
+  };
+
+  next();
 };
 
-module.exports = {
-    upload,
-    uploadSingle,
-    uploadMultiple,
-    uploadFields,
-    cleanupUploads,
-    uploadDir
+export default {
+  upload,
+  uploadSingle,
+  uploadMultiple,
+  uploadFields,
+  cleanupUploads,
+  uploadDir,
 };
