@@ -1,33 +1,33 @@
-import express from 'express';
-import cors from 'cors';
-import helmet from 'helmet';
-import morgan from 'morgan';
-import compression from 'compression';
-import cookieParser from 'cookie-parser';
-import dotenv from 'dotenv';
+import express from "express";
+import cors from "cors";
+import helmet from "helmet";
+import morgan from "morgan";
+import compression from "compression";
+import cookieParser from "cookie-parser";
+import dotenv from "dotenv";
 
 // Import middleware
-import errorMiddleware from './middleware/errorMiddleware.js';
-import loggerMiddleware from './middleware/loggerMiddleware.js';
-import sanitizeMiddleware from './middleware/sanitizeMiddleware.js';
-import rateLimiter from './middleware/rateLimiter.js';
-import { configureCors } from './config/cors.js';
+import errorMiddleware from "./middleware/errorMiddleware.js";
+import loggerMiddleware from "./middleware/loggerMiddleware.js";
+import sanitizeMiddleware from "./middleware/sanitizeMiddleware.js";
+import rateLimiter from "./middleware/rateLimiter.js";
+import { configureCors } from "./config/cors.js";
 
 // Import routes
-import authRoutes from './routes/authRoutes.js';
-import userRoutes from './routes/userRoutes.js';
-import donationRoutes from './routes/donationRoutes.js';
-import adminRoutes from './routes/adminRoutes.js';
-import volunteerRoutes from './routes/volunteerRoutes.js';
-import fundingRoutes from './routes/fundingRoutes.js';
-import searchRoutes from './routes/searchRoutes.js';
-import contactRoutes from './routes/contactRoutes.js';
-import dashboardRoutes from './routes/dashboardRoutes.js';
-import analyticsRoutes from './routes/analyticsRoutes.js';
-import indexRoutes from './routes/index.js';
+import authRoutes from "./routes/authRoutes.js";
+import userRoutes from "./routes/userRoutes.js";
+import donationRoutes from "./routes/donationRoutes.js";
+import adminRoutes from "./routes/adminRoutes.js";
+import volunteerRoutes from "./routes/volunteerRoutes.js";
+import fundingRoutes from "./routes/fundingRoutes.js";
+import searchRoutes from "./routes/searchRoutes.js";
+import contactRoutes from "./routes/contactRoutes.js";
+import dashboardRoutes from "./routes/dashboardRoutes.js";
+import analyticsRoutes from "./routes/analyticsRoutes.js";
+import indexRoutes from "./routes/index.js";
 
 // Import database connection
-import connectDB from './config/db.js';
+import connectDB from "./config/db.js";
 
 // Load environment variables
 dotenv.config();
@@ -39,44 +39,46 @@ const app = express();
 connectDB();
 
 // Apply security middleware
-app.use(helmet({
-  contentSecurityPolicy: {
-    directives: {
-      defaultSrc: ["'self'"],
-      styleSrc: ["'self'", "'unsafe-inline'", "https://fonts.googleapis.com"],
-      scriptSrc: ["'self'", "'unsafe-inline'", "https://js.stripe.com"],
-      imgSrc: ["'self'", "data:", "https:", "http:"],
-      connectSrc: ["'self'", "https://api.stripe.com"],
-      fontSrc: ["'self'", "https://fonts.gstatic.com"],
+app.use(
+  helmet({
+    contentSecurityPolicy: {
+      directives: {
+        defaultSrc: ["'self'"],
+        styleSrc: ["'self'", "'unsafe-inline'", "https://fonts.googleapis.com"],
+        scriptSrc: ["'self'", "'unsafe-inline'", "https://js.stripe.com"],
+        imgSrc: ["'self'", "data:", "https:", "http:"],
+        connectSrc: ["'self'", "https://api.stripe.com"],
+        fontSrc: ["'self'", "https://fonts.gstatic.com"],
+      },
     },
-  },
-  crossOriginEmbedderPolicy: false,
-}));
+    crossOriginEmbedderPolicy: false,
+  })
+);
 
 // Configure CORS
 app.use(configureCors());
 
 // Apply rate limiting
-app.use('/api/', rateLimiter);
+app.use("/api/", rateLimiter.apiLimiter);
 
 // Other middleware
 app.use(compression());
-app.use(express.json({ limit: '10mb' }));
-app.use(express.urlencoded({ extended: true, limit: '10mb' }));
+app.use(express.json({ limit: "10mb" }));
+app.use(express.urlencoded({ extended: true, limit: "10mb" }));
 app.use(cookieParser());
-app.use(sanitizeMiddleware);
+app.use(sanitizeMiddleware.xssProtect);
 
 // Logger middleware (only in development)
-if (process.env.NODE_ENV === 'development') {
-  app.use(morgan('dev'));
-  app.use(loggerMiddleware);
+if (process.env.NODE_ENV === "development") {
+  app.use(morgan("dev"));
+  app.use(loggerMiddleware.httpLogger);
 }
 
 // Health check route
-app.get('/health', (req, res) => {
+app.get("/health", (req, res) => {
   res.status(200).json({
-    status: 'success',
-    message: 'Blood Donation API is running',
+    status: "success",
+    message: "Blood Donation API is running",
     timestamp: new Date().toISOString(),
     uptime: process.uptime(),
     environment: process.env.NODE_ENV,
@@ -84,39 +86,39 @@ app.get('/health', (req, res) => {
 });
 
 // API routes
-app.use('/api', indexRoutes);
-app.use('/api/auth', authRoutes);
-app.use('/api/users', userRoutes);
-app.use('/api/donations', donationRoutes);
-app.use('/api/admin', adminRoutes);
-app.use('/api/volunteer', volunteerRoutes);
-app.use('/api/funding', fundingRoutes);
-app.use('/api/search', searchRoutes);
-app.use('/api/contact', contactRoutes);
-app.use('/api/dashboard', dashboardRoutes);
-app.use('/api/analytics', analyticsRoutes);
+app.use("/api", indexRoutes);
+app.use("/api/auth", authRoutes);
+app.use("/api/users", userRoutes);
+app.use("/api/donations", donationRoutes);
+app.use("/api/admin", adminRoutes);
+app.use("/api/volunteer", volunteerRoutes);
+app.use("/api/funding", fundingRoutes);
+app.use("/api/search", searchRoutes);
+app.use("/api/contact", contactRoutes);
+app.use("/api/dashboard", dashboardRoutes);
+app.use("/api/analytics", analyticsRoutes);
 
 // Serve static files in production
-if (process.env.NODE_ENV === 'production') {
-  import('path').then((path) => {
+if (process.env.NODE_ENV === "production") {
+  import("path").then((path) => {
     const __dirname = path.resolve();
-    app.use(express.static(path.join(__dirname, '../client/dist')));
-    
-    app.get('*', (req, res) => {
-      res.sendFile(path.join(__dirname, '../client/dist/index.html'));
+    app.use(express.static(path.join(__dirname, "../client/dist")));
+
+    app.get("*", (req, res) => {
+      res.sendFile(path.join(__dirname, "../client/dist/index.html"));
     });
   });
 }
 
 // 404 handler
-app.use('*', (req, res) => {
+app.use("*", (req, res) => {
   res.status(404).json({
-    status: 'error',
+    status: "error",
     message: `Route ${req.originalUrl} not found`,
   });
 });
 
 // Global error handler
-app.use(errorMiddleware);
+app.use(errorMiddleware.errorHandler);
 
 export default app;
