@@ -1,58 +1,63 @@
-import jwt from 'jsonwebtoken';
-import User from '../models/User.js';
+import jwt from "jsonwebtoken";
+import User from "../models/User.js";
 
 const protect = async (req, res, next) => {
-    let token;
+  let token;
 
-    if (
-        req.headers.authorization &&
-        req.headers.authorization.startsWith('Bearer')
-    ) {
-        try {
-            token = req.headers.authorization.split(' ')[1];
-            const decoded = jwt.verify(token, process.env.JWT_SECRET);
-            req.user = await User.findById(decoded.id).select('-password');
-            
-            if (!req.user) {
-                return res.status(401).json({
-                    success: false,
-                    message: 'User not found'
-                });
-            }
-            
-            next();
-        } catch (error) {
-            console.error('Auth middleware error:', error);
-            return res.status(401).json({
-                success: false,
-                message: 'Not authorized, token failed'
-            });
-        }
-    }
+  if (
+    req.headers.authorization &&
+    req.headers.authorization.startsWith("Bearer")
+  ) {
+    try {
+      token = req.headers.authorization.split(" ")[1];
+      const decoded = jwt.verify(token, process.env.JWT_SECRET);
+      req.user = await User.findById(decoded.id).select("-password");
 
-    if (!token) {
+      if (!req.user) {
         return res.status(401).json({
-            success: false,
-            message: 'Not authorized, no token'
+          success: false,
+          message: "User not found",
         });
+      }
+
+      next();
+    } catch (error) {
+      console.error("Auth middleware error:", error);
+      return res.status(401).json({
+        success: false,
+        message: "Not authorized, token failed",
+      });
     }
+  }
+
+  if (!token) {
+    return res.status(401).json({
+      success: false,
+      message: "Not authorized, no token",
+    });
+  }
 };
 
 const optionalAuth = async (req, res, next) => {
-    if (
-        req.headers.authorization &&
-        req.headers.authorization.startsWith('Bearer')
-    ) {
-        try {
-            const token = req.headers.authorization.split(' ')[1];
-            const decoded = jwt.verify(token, process.env.JWT_SECRET);
-            req.user = await User.findById(decoded.id).select('-password');
-        } catch (error) {
-            console.error('Optional auth error:', error);
-        }
+  if (
+    req.headers.authorization &&
+    req.headers.authorization.startsWith("Bearer")
+  ) {
+    try {
+      const token = req.headers.authorization.split(" ")[1];
+      const decoded = jwt.verify(token, process.env.JWT_SECRET);
+      req.user = await User.findById(decoded.id).select("-password");
+    } catch (error) {
+      console.error("Optional auth error:", error);
     }
-    next();
+  }
+  next();
 };
 
 // Export as ES6
 export { protect, optionalAuth };
+
+export default {
+  protect,
+  optionalAuth,
+};
